Calculate clonotype abundance

Calculate clonotype abundance

Usage

calc_abundance(
  input,
  cluster_col = NULL,
  clonotype_col = "clonotype_id",
  prefix = "",
  return_df = FALSE
)

Arguments

input

Single cell object or data.frame containing V(D)J data. If a data.frame is provided, the cell barcodes should be stored as row names.

cluster_col

meta.data column containing cluster IDs to use for grouping cells when calculating clonotype abundance

clonotype_col

meta.data column containing clonotype IDs to use for calculating clonotype abundance

prefix

Prefix to add to new columns

return_df

Return results as a data.frame. If set to FALSE, results will be added to the input object.

Value

Single cell object or data.frame with clonotype abundance metrics

Examples

# Calculate clonotype abundance using all cells
res <- calc_abundance(
  vdj_so,
  clonotype_col = "clonotype_id"
)

head(res@meta.data, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAACCTGAGTAGGTGC-1     avid_1          1            1               1
#>                      seurat_clusters   UMAP_1     UMAP_2 clonotype_id   v_gene
#> 1_AAACCTGAGTAGGTGC-1               1 2.561658 -0.6746224   clonotype1 IGKV5-43
#>                      d_gene j_gene c_gene chains isotype        cdr3
#> 1_AAACCTGAGTAGGTGC-1   None  IGKJ2   IGKC    IGK    None CQQSNSWPYTF
#>                                                cdr3_nt reads umis productive
#> 1_AAACCTGAGTAGGTGC-1 TGTCAACAGAGTAACAGCTGGCCGTACACGTTC  1054   70       TRUE
#>                      full_length paired n_insertion n_deletion n_mismatch
#> 1_AAACCTGAGTAGGTGC-1        TRUE  FALSE           0          4          2
#>                      cdr3_length cdr3_nt_length n_chains clone_freq clone_pct
#> 1_AAACCTGAGTAGGTGC-1          11             33        1         38  42.69663

# Group cells based on meta.data column before calculating abundance
res <- calc_abundance(
  vdj_sce,
  cluster_col = "orig.ident"
)

head(res@colData, 1)
#> DataFrame with 1 row and 30 columns
#>                       orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#>                      <character>  <numeric>    <integer>        <factor>
#> 1_AAACCTGAGTAGGTGC-1      avid_1          1            1               1
#>                      seurat_clusters    UMAP_1    UMAP_2 clonotype_id
#>                             <factor> <numeric> <numeric>  <character>
#> 1_AAACCTGAGTAGGTGC-1               1   2.56166 -0.674622   clonotype1
#>                           v_gene      d_gene      j_gene      c_gene
#>                      <character> <character> <character> <character>
#> 1_AAACCTGAGTAGGTGC-1    IGKV5-43        None       IGKJ2        IGKC
#>                           chains     isotype        cdr3                cdr3_nt
#>                      <character> <character> <character>            <character>
#> 1_AAACCTGAGTAGGTGC-1         IGK        None CQQSNSWPYTF TGTCAACAGAGTAACAGCTG..
#>                            reads        umis  productive full_length    paired
#>                      <character> <character> <character> <character> <logical>
#> 1_AAACCTGAGTAGGTGC-1        1054          70        TRUE        TRUE     FALSE
#>                      n_insertion  n_deletion  n_mismatch cdr3_length
#>                      <character> <character> <character> <character>
#> 1_AAACCTGAGTAGGTGC-1           0           4           2          11
#>                      cdr3_nt_length  n_chains clone_freq clone_pct clone_shared
#>                         <character> <integer>  <integer> <numeric>    <logical>
#> 1_AAACCTGAGTAGGTGC-1             33         1         18        36         TRUE

# Add a prefix to the new columns
# this is useful if multiple abundance calculations are stored in the
# meta.data
res <- calc_abundance(
  vdj_so,
  prefix = "bcr_"
)

head(res@meta.data, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAACCTGAGTAGGTGC-1     avid_1          1            1               1
#>                      seurat_clusters   UMAP_1     UMAP_2 clonotype_id   v_gene
#> 1_AAACCTGAGTAGGTGC-1               1 2.561658 -0.6746224   clonotype1 IGKV5-43
#>                      d_gene j_gene c_gene chains isotype        cdr3
#> 1_AAACCTGAGTAGGTGC-1   None  IGKJ2   IGKC    IGK    None CQQSNSWPYTF
#>                                                cdr3_nt reads umis productive
#> 1_AAACCTGAGTAGGTGC-1 TGTCAACAGAGTAACAGCTGGCCGTACACGTTC  1054   70       TRUE
#>                      full_length paired n_insertion n_deletion n_mismatch
#> 1_AAACCTGAGTAGGTGC-1        TRUE  FALSE           0          4          2
#>                      cdr3_length cdr3_nt_length n_chains bcr_clone_freq
#> 1_AAACCTGAGTAGGTGC-1          11             33        1             38
#>                      bcr_clone_pct
#> 1_AAACCTGAGTAGGTGC-1      42.69663

# Return a data.frame instead of adding the results to the input object
res <- calc_abundance(
  vdj_sce,
  return_df = TRUE
)

head(res, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAACCTGAGTAGGTGC-1     avid_1          1            1               1
#>                      seurat_clusters   UMAP_1     UMAP_2 clonotype_id   v_gene
#> 1_AAACCTGAGTAGGTGC-1               1 2.561658 -0.6746224   clonotype1 IGKV5-43
#>                      d_gene j_gene c_gene chains isotype        cdr3
#> 1_AAACCTGAGTAGGTGC-1   None  IGKJ2   IGKC    IGK    None CQQSNSWPYTF
#>                                                cdr3_nt reads umis productive
#> 1_AAACCTGAGTAGGTGC-1 TGTCAACAGAGTAACAGCTGGCCGTACACGTTC  1054   70       TRUE
#>                      full_length paired n_insertion n_deletion n_mismatch
#> 1_AAACCTGAGTAGGTGC-1        TRUE  FALSE           0          4          2
#>                      cdr3_length cdr3_nt_length n_chains clone_freq clone_pct
#> 1_AAACCTGAGTAGGTGC-1          11             33        1         38  42.69663