Calculate V(D)J segment usage

Calculate V(D)J segment usage

Usage

calc_gene_usage(
  input,
  data_cols,
  cluster_col = NULL,
  chain = NULL,
  chain_col = "chains",
  sep = ";"
)

Arguments

input

Object containing V(D)J data. If a data.frame is provided, the cell barcodes should be stored as row names.

data_cols

meta.data column containing V(D)J genes identified for each clonotype. If multiple columns are provided, paired usage of genes will be calculated.

cluster_col

meta.data column containing cell clusters to use when calculating gene usage

chain

Chain to use for calculating gene usage. Set to NULL to include all chains.

chain_col

meta.data column containing chains for each cell

sep

Separator used for storing per cell V(D)J data

Value

data.frame containing gene usage summary

See also

Examples

# Calculate V(D)J segment usage for all cells
calc_gene_usage(
  vdj_so,
  data_cols = "v_gene"
)
#> # A tibble: 62 × 4
#>    v_gene     n_cells  freq   pct
#>    <chr>        <int> <int> <dbl>
#>  1 IGKV5-43        89    53 59.6 
#>  2 IGKV10-96       89     4  4.49
#>  3 IGHV1-82        89     3  3.37
#>  4 IGKV1-117       89     3  3.37
#>  5 IGKV10-94       89     3  3.37
#>  6 IGKV14-111      89     3  3.37
#>  7 IGKV2-137       89     3  3.37
#>  8 IGKV3-4         89     3  3.37
#>  9 IGKV6-15        89     3  3.37
#> 10 IGKV1-110       89     2  2.25
#> # … with 52 more rows

# Calculate gene usage separately for cell clusters
calc_gene_usage(
  vdj_sce,
  data_cols = "v_gene",
  cluster_col = "orig.ident"
)
#> # A tibble: 124 × 5
#>    v_gene    orig.ident n_cells  freq   pct
#>    <chr>     <chr>        <dbl> <int> <dbl>
#>  1 IGKV5-43  avid_2          39    26 66.7 
#>  2 IGKV5-43  avid_1          50    27 54   
#>  3 IGKV1-117 avid_1          50     3  6   
#>  4 IGHV1-82  avid_2          39     2  5.13
#>  5 IGKV1-135 avid_2          39     2  5.13
#>  6 IGKV10-96 avid_2          39     2  5.13
#>  7 IGKV2-137 avid_2          39     2  5.13
#>  8 IGKV6-15  avid_2          39     2  5.13
#>  9 IGKV1-110 avid_1          50     2  4   
#> 10 IGKV10-94 avid_1          50     2  4   
#> # … with 114 more rows

# Calculate gene usage for a specific chain(s)
calc_gene_usage(
  vdj_so,
  data_cols = "v_gene",
  chain = c("IGK", "IGL")
)
#> # A tibble: 38 × 4
#>    v_gene     n_cells  freq   pct
#>    <chr>        <int> <int> <dbl>
#>  1 IGKV5-43        89    53 59.6 
#>  2 IGKV10-96       89     4  4.49
#>  3 IGKV1-117       89     3  3.37
#>  4 IGKV10-94       89     3  3.37
#>  5 IGKV14-111      89     3  3.37
#>  6 IGKV2-137       89     3  3.37
#>  7 IGKV3-4         89     3  3.37
#>  8 IGKV6-15        89     3  3.37
#>  9 IGKV1-110       89     2  2.25
#> 10 IGKV1-135       89     2  2.25
#> # … with 28 more rows

# Calculate paired usage of V(D)J segments
calc_gene_usage(
  vdj_sce,
  data_cols = c("v_gene", "j_gene"),
)
#> # A tibble: 76 × 5
#>    v_gene    j_gene n_cells  freq   pct
#>    <chr>     <chr>    <int> <int> <dbl>
#>  1 IGKV5-43  IGKJ2       89    53 59.6 
#>  2 IGKV10-96 IGKJ1       89     3  3.37
#>  3 IGKV2-137 IGKJ2       89     3  3.37
#>  4 IGHV1-82  IGHJ4       89     2  2.25
#>  5 IGKV1-117 IGKJ2       89     2  2.25
#>  6 IGKV10-94 IGKJ1       89     2  2.25
#>  7 IGKV3-4   IGKJ2       89     2  2.25
#>  8 IGKV6-15  IGKJ2       89     2  2.25
#>  9 IGHV1-18  IGHJ4       89     1  1.12
#> 10 IGHV1-22  IGHJ3       89     1  1.12
#> # … with 66 more rows