Calculate V(D)J gene segment usage

Calculate V(D)J gene segment usage

calc_vdj_usage(
  input,
  gene_cols,
  cluster_col = NULL,
  chain = NULL,
  chain_col = "chains",
  sep = ";"
)

Arguments

input

Object containing V(D)J data. If a data.frame is provided, the cell barcodes should be stored as row names.

gene_cols

meta.data column containing V(D)J genes identified for each clonotype. If multiple columns are provided, paired usage of genes will be calculated.

cluster_col

meta.data column containing cell clusters to use when calculating gene usage

chain

Chain to use for calculating gene usage. Set to NULL to include all chains.

chain_col

meta.data column containing chains for each cell

sep

Separator used for storing per cell V(D)J data

Value

data.frame containing gene usage summary

Examples

# Calculate V(D)J segment usage for all cells
calc_vdj_usage(
  vdj_so,
  gene_cols = "v_gene"
)
#> # A tibble: 62 × 4
#>    v_gene     n_cells  freq   pct
#>    <chr>        <int> <int> <dbl>
#>  1 IGHV1-18        89     1  1.12
#>  2 IGHV1-22        89     1  1.12
#>  3 IGHV1-26        89     1  1.12
#>  4 IGHV1-47        89     1  1.12
#>  5 IGHV1-52        89     1  1.12
#>  6 IGHV1-53        89     1  1.12
#>  7 IGHV1-55        89     1  1.12
#>  8 IGHV1-62-2      89     1  1.12
#>  9 IGHV1-64        89     1  1.12
#> 10 IGHV1-66        89     1  1.12
#> # … with 52 more rows

# Calculate gene usage separately for cell clusters
calc_vdj_usage(
  vdj_so,
  gene_cols = "v_gene",
  cluster_col = "orig.ident"
)
#> # A tibble: 124 × 5
#>    v_gene   orig.ident n_cells  freq   pct
#>    <chr>    <chr>        <dbl> <int> <dbl>
#>  1 IGHV1-18 avid_1          50     1  2   
#>  2 IGHV1-18 avid_2          39     0  0   
#>  3 IGHV1-22 avid_1          50     0  0   
#>  4 IGHV1-22 avid_2          39     1  2.56
#>  5 IGHV1-26 avid_1          50     1  2   
#>  6 IGHV1-26 avid_2          39     0  0   
#>  7 IGHV1-47 avid_1          50     0  0   
#>  8 IGHV1-47 avid_2          39     1  2.56
#>  9 IGHV1-52 avid_1          50     0  0   
#> 10 IGHV1-52 avid_2          39     1  2.56
#> # … with 114 more rows

# Calculate gene usage for a specific chain
calc_vdj_usage(
  vdj_so,
  gene_cols = "v_gene",
  chain = c("IGH", "IGK")
)
#> # A tibble: 60 × 4
#>    v_gene     n_cells  freq   pct
#>    <chr>        <int> <int> <dbl>
#>  1 IGHV1-18        89     1  1.12
#>  2 IGHV1-22        89     1  1.12
#>  3 IGHV1-26        89     1  1.12
#>  4 IGHV1-47        89     1  1.12
#>  5 IGHV1-52        89     1  1.12
#>  6 IGHV1-53        89     1  1.12
#>  7 IGHV1-55        89     1  1.12
#>  8 IGHV1-62-2      89     1  1.12
#>  9 IGHV1-64        89     1  1.12
#> 10 IGHV1-66        89     1  1.12
#> # … with 50 more rows

# Calculate paired usage of V(D)J segments
calc_vdj_usage(
  vdj_so,
  gene_cols = c("v_gene", "j_gene"),
)
#> # A tibble: 76 × 5
#>    v_gene     j_gene n_cells  freq   pct
#>    <chr>      <chr>    <int> <int> <dbl>
#>  1 IGHV1-18   IGHJ4       89     1  1.12
#>  2 IGHV1-22   IGHJ3       89     1  1.12
#>  3 IGHV1-26   IGHJ4       89     1  1.12
#>  4 IGHV1-47   IGHJ3       89     1  1.12
#>  5 IGHV1-52   IGHJ2       89     1  1.12
#>  6 IGHV1-53   IGHJ3       89     1  1.12
#>  7 IGHV1-55   IGHJ3       89     1  1.12
#>  8 IGHV1-62-2 IGHJ1       89     1  1.12
#>  9 IGHV1-64   IGHJ4       89     1  1.12
#> 10 IGHV1-66   IGHJ2       89     1  1.12
#> # … with 66 more rows

# Using calc_vdj_usage outside of Seurat
# SingleCellExperiment objects or data.frames containing V(D)J data are also
# compatible. If a data.frame is provided, cell barcodes should be stored as
# row names.
calc_vdj_usage(vdj_sce, gene_cols = "v_gene")
#> # A tibble: 62 × 4
#>    v_gene     n_cells  freq   pct
#>    <chr>        <int> <int> <dbl>
#>  1 IGHV1-18        89     1  1.12
#>  2 IGHV1-22        89     1  1.12
#>  3 IGHV1-26        89     1  1.12
#>  4 IGHV1-47        89     1  1.12
#>  5 IGHV1-52        89     1  1.12
#>  6 IGHV1-53        89     1  1.12
#>  7 IGHV1-55        89     1  1.12
#>  8 IGHV1-62-2      89     1  1.12
#>  9 IGHV1-64        89     1  1.12
#> 10 IGHV1-66        89     1  1.12
#> # … with 52 more rows

df <- vdj_so@meta.data

calc_vdj_usage(df, gene_cols = "v_gene")
#> # A tibble: 62 × 4
#>    v_gene     n_cells  freq   pct
#>    <chr>        <int> <int> <dbl>
#>  1 IGHV1-18        89     1  1.12
#>  2 IGHV1-22        89     1  1.12
#>  3 IGHV1-26        89     1  1.12
#>  4 IGHV1-47        89     1  1.12
#>  5 IGHV1-52        89     1  1.12
#>  6 IGHV1-53        89     1  1.12
#>  7 IGHV1-55        89     1  1.12
#>  8 IGHV1-62-2      89     1  1.12
#>  9 IGHV1-64        89     1  1.12
#> 10 IGHV1-66        89     1  1.12
#> # … with 52 more rows