Fetch V(D)J data

Fetch per-chain V(D)J data from object. Within the object meta.data, each row represents a single cell and can include information for multiple chains. This function returns an unnested data.frame where each row represents a single chain. This is useful for plotting per-chain metrics such as CDR3 length or the number of insertions/deletions.

fetch_vdj(
  input,
  vdj_cols = NULL,
  clonotype_col = NULL,
  filter_cells = FALSE,
  unnest = TRUE,
  sep = ";"
)

Arguments

input

Single cell object or data.frame containing V(D)J data. If a data.frame is provided, the cell barcodes should be stored as row names.

vdj_cols

meta.data columns containing per-chain V(D)J data to unnest. If NULL, V(D)J data are automatically selected by identifying columns that have NAs in the same rows as clonotype_col.

clonotype_col

meta.data column containing clonotype IDs. This column is used to determine which columns have V(D)J data. If both clonotype_col and vdj_cols are NULL, all columns are included.

filter_cells

Remove cells that do not have V(D)J data, clonotype_col must be provided to determine which cells to filter.

unnest

If FALSE, a nested data.frame is returned where each row represents a cell and V(D)J data is stored as list-cols. If TRUE, columns are unnested so each row represents a single chain.

sep

Separator used for storing per cell V(D)J data. This is used to identify columns containing per-chain data that can be unnested.

Value

data.frame containing V(D)J data

Examples

# Fetch per-chain V(D)J data
fetch_vdj(vdj_so)
#> # A tibble: 258 × 29
#>    .cell_id      orig.ident nCount_RNA nFeature_RNA Percent_mito RNA_snn_res.0.3
#>    <chr>         <chr>           <dbl>        <int>        <dbl> <fct>          
#>  1 1_AAACCTGAGT… avid_1              1            1            0 1              
#>  2 1_AAACCTGAGT… avid_1              0            0          NaN 1              
#>  3 1_AAACCTGAGT… avid_1              0            0          NaN 1              
#>  4 1_AAACCTGCAG… avid_1              4            4            0 0              
#>  5 1_AAACCTGCAG… avid_1              4            4            0 0              
#>  6 1_AAACCTGCAG… avid_1              4            4            0 0              
#>  7 1_AAACCTGCAT… avid_1              9            8            0 0              
#>  8 1_AACGTTGTCG… avid_1              1            1            0 1              
#>  9 1_AATCGGTAGT… avid_1              1            1            0 1              
#> 10 1_ACATCAGAGT… avid_1              2            2            0 0              
#> # … with 248 more rows, and 23 more variables: seurat_clusters <fct>,
#> #   UMAP_1 <dbl>, UMAP_2 <dbl>, clonotype_id <chr>, v_gene <chr>, d_gene <chr>,
#> #   j_gene <chr>, c_gene <chr>, chains <chr>, isotype <chr>, cdr3 <chr>,
#> #   cdr3_nt <chr>, reads <dbl>, umis <dbl>, productive <lgl>,
#> #   full_length <lgl>, paired <lgl>, n_insertion <dbl>, n_deletion <dbl>,
#> #   n_mismatch <dbl>, cdr3_length <dbl>, cdr3_nt_length <dbl>, n_chains <int>

# To increase performance, specify which columns to return per-chain data,
# per-cell data will be returned for all other columns
fetch_vdj(
  vdj_sce,
  vdj_cols = c("chains", "n_insertion")
)
#> # A tibble: 258 × 29
#>    .cell_id      orig.ident nCount_RNA nFeature_RNA Percent_mito RNA_snn_res.0.3
#>    <chr>         <chr>           <dbl>        <int>        <dbl> <fct>          
#>  1 1_AAACCTGAGT… avid_1              1            1            0 1              
#>  2 1_AAACCTGAGT… avid_1              0            0          NaN 1              
#>  3 1_AAACCTGAGT… avid_1              0            0          NaN 1              
#>  4 1_AAACCTGCAG… avid_1              4            4            0 0              
#>  5 1_AAACCTGCAG… avid_1              4            4            0 0              
#>  6 1_AAACCTGCAG… avid_1              4            4            0 0              
#>  7 1_AAACCTGCAT… avid_1              9            8            0 0              
#>  8 1_AACGTTGTCG… avid_1              1            1            0 1              
#>  9 1_AATCGGTAGT… avid_1              1            1            0 1              
#> 10 1_ACATCAGAGT… avid_1              2            2            0 0              
#> # … with 248 more rows, and 23 more variables: seurat_clusters <fct>,
#> #   UMAP_1 <dbl>, UMAP_2 <dbl>, clonotype_id <chr>, v_gene <chr>, d_gene <chr>,
#> #   j_gene <chr>, c_gene <chr>, chains <chr>, isotype <chr>, cdr3 <chr>,
#> #   cdr3_nt <chr>, reads <chr>, umis <chr>, productive <chr>,
#> #   full_length <chr>, paired <lgl>, n_insertion <dbl>, n_deletion <chr>,
#> #   n_mismatch <chr>, cdr3_length <chr>, cdr3_nt_length <chr>, n_chains <int>

# Only include cells that have V(D)J data
# clonotype_col must be specified to identify cells with V(D)J data
fetch_vdj(
  vdj_so,
  filter_cells = TRUE,
  clonotype_col = "clonotype_id"
)
#> # A tibble: 143 × 29
#>    .cell_id      orig.ident nCount_RNA nFeature_RNA Percent_mito RNA_snn_res.0.3
#>    <chr>         <chr>           <dbl>        <int>        <dbl> <fct>          
#>  1 1_AAACCTGAGT… avid_1              1            1            0 1              
#>  2 1_AAACCTGAGT… avid_1              0            0          NaN 1              
#>  3 1_AAACCTGAGT… avid_1              0            0          NaN 1              
#>  4 1_AAACCTGCAG… avid_1              4            4            0 0              
#>  5 1_AAACCTGCAG… avid_1              4            4            0 0              
#>  6 1_AAACCTGCAG… avid_1              4            4            0 0              
#>  7 1_AAACCTGCAT… avid_1              9            8            0 0              
#>  8 1_ACCTTTATCA… avid_1              4            4            0 0              
#>  9 1_ACCTTTATCA… avid_1              4            4            0 0              
#> 10 1_ACCTTTATCA… avid_1              4            4            0 0              
#> # … with 133 more rows, and 23 more variables: seurat_clusters <fct>,
#> #   UMAP_1 <dbl>, UMAP_2 <dbl>, clonotype_id <chr>, v_gene <chr>, d_gene <chr>,
#> #   j_gene <chr>, c_gene <chr>, chains <chr>, isotype <chr>, cdr3 <chr>,
#> #   cdr3_nt <chr>, reads <dbl>, umis <dbl>, productive <lgl>,
#> #   full_length <lgl>, paired <lgl>, n_insertion <dbl>, n_deletion <dbl>,
#> #   n_mismatch <dbl>, cdr3_length <dbl>, cdr3_nt_length <dbl>, n_chains <int>