Filter V(D)J data in object

Remove V(D)J data for chains/cells that do not satisfy the provided condition

Usage

filter_vdj(
  input,
  filt,
  data_cols = NULL,
  clonotype_col = "clonotype_id",
  sep = ";",
  per_cell = FALSE
)

Arguments

input

Single cell object or data.frame containing V(D)J data. If a data.frame is provided, cell barcodes should be stored as row names.

filt

Condition to use for filtering V(D)J data. To allow for filtering of per-chain V(D)J data, the data for each cell is converted into a vector, e.g. 'IGH;IGK' is equivalent to c('IGH', 'IGK'). This allows R vector operations to be performed on the per-chain values. The filtering condition must return TRUE/FALSE for each chain or a single TRUE/FALSE for each cell. Data can be filtered based on cell barcodes by referring to the '.cell_id' column.

data_cols

meta.data columns containing V(D)J data to use for filtering. If NULL, V(D)J data are automatically selected by identifying columns that have NAs in the same rows as clonotype_col.

clonotype_col

meta.data column containing clonotype IDs. This column is used to determine which columns contain V(D)J data.

sep

Separator used for storing per cell V(D)J data. If NULL, per_cell will be set TRUE.

per_cell

If TRUE, per-chain data will not be parsed and values in each meta.data column will be filtered as is.

Value

Object with filtered meta.data

Examples

# Only include V(D)J data for productive chains
res <- filter_vdj(vdj_so, productive)

head(res@meta.data, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAACCTGAGTAGGTGC-1     avid_1          1            1               1
#>                      seurat_clusters   UMAP_1     UMAP_2 clonotype_id
#> 1_AAACCTGAGTAGGTGC-1               1 2.561658 -0.6746224   clonotype1
#>                        v_gene d_gene j_gene c_gene chains isotype
#> 1_AAACCTGAGTAGGTGC-1 IGKV5-43   None  IGKJ2   IGKC    IGK    None
#>                             cdr3                           cdr3_nt reads
#> 1_AAACCTGAGTAGGTGC-1 CQQSNSWPYTF TGTCAACAGAGTAACAGCTGGCCGTACACGTTC  1054
#>                      umis productive full_length paired n_insertion
#> 1_AAACCTGAGTAGGTGC-1   70       TRUE        TRUE  FALSE           0
#>                      n_deletion n_mismatch cdr3_length cdr3_nt_length
#> 1_AAACCTGAGTAGGTGC-1          4          2          11             33
#>                      n_chains
#> 1_AAACCTGAGTAGGTGC-1        1

# Only include V(D)J data for cells with paired chains
res <- filter_vdj(vdj_sce, paired)

head(res@colData, 3)
#> DataFrame with 3 rows and 27 columns
#>                       orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#>                      <character>  <numeric>    <integer>        <factor>
#> 1_AAACCTGAGTAGGTGC-1      avid_1          1            1               1
#> 1_AAACCTGAGTCAAGCG-1      avid_1          0            0               1
#> 1_AAACCTGCAGTTTACG-1      avid_1          4            4               0
#>                      seurat_clusters    UMAP_1    UMAP_2 clonotype_id
#>                             <factor> <numeric> <numeric>  <character>
#> 1_AAACCTGAGTAGGTGC-1               1   2.56166 -0.674622           NA
#> 1_AAACCTGAGTCAAGCG-1               1  -3.40268  6.838597           NA
#> 1_AAACCTGCAGTTTACG-1               0   1.19011 -2.108460 clonotype537
#>                                      v_gene         d_gene
#>                                 <character>    <character>
#> 1_AAACCTGAGTAGGTGC-1                     NA             NA
#> 1_AAACCTGAGTCAAGCG-1                     NA             NA
#> 1_AAACCTGCAGTTTACG-1 IGHV1-9;IGKV5-43;IGK.. None;None;None
#>                                 j_gene         c_gene      chains
#>                            <character>    <character> <character>
#> 1_AAACCTGAGTAGGTGC-1                NA             NA          NA
#> 1_AAACCTGAGTCAAGCG-1                NA             NA          NA
#> 1_AAACCTGCAGTTTACG-1 IGHJ4;IGKJ2;IGKJ1 IGHM;IGKC;IGKC IGH;IGK;IGK
#>                          isotype                   cdr3
#>                      <character>            <character>
#> 1_AAACCTGAGTAGGTGC-1          NA                     NA
#> 1_AAACCTGAGTCAAGCG-1          NA                     NA
#> 1_AAACCTGCAGTTTACG-1        IGHM CATLWYDAMDYW;CQQSNSW..
#>                                     cdr3_nt       reads        umis
#>                                 <character> <character> <character>
#> 1_AAACCTGAGTAGGTGC-1                     NA          NA          NA
#> 1_AAACCTGAGTCAAGCG-1                     NA          NA          NA
#> 1_AAACCTGCAGTTTACG-1 TGTGCAACTCTATGGTACGA.. 144;460;156     7;28;10
#>                          productive    full_length    paired n_insertion
#>                         <character>    <character> <logical> <character>
#> 1_AAACCTGAGTAGGTGC-1             NA             NA        NA          NA
#> 1_AAACCTGAGTCAAGCG-1             NA             NA        NA          NA
#> 1_AAACCTGCAGTTTACG-1 TRUE;TRUE;TRUE TRUE;TRUE;TRUE      TRUE       4;0;0
#>                       n_deletion  n_mismatch cdr3_length cdr3_nt_length
#>                      <character> <character> <character>    <character>
#> 1_AAACCTGAGTAGGTGC-1          NA          NA          NA             NA
#> 1_AAACCTGAGTCAAGCG-1          NA          NA          NA             NA
#> 1_AAACCTGCAGTTTACG-1       0;4;3       4;2;0    12;11;11       36;33;33
#>                       n_chains
#>                      <integer>
#> 1_AAACCTGAGTAGGTGC-1        NA
#> 1_AAACCTGAGTCAAGCG-1        NA
#> 1_AAACCTGCAGTTTACG-1         3

# Only include V(D)J data for cells with at least one heavy and one light chain
res <- filter_vdj(
  vdj_so,
  "IGH" %in% chains && any(c("IGK", "IGL") %in% chains)
)

head(res@meta.data, 3)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAACCTGAGTAGGTGC-1     avid_1          1            1               1
#> 1_AAACCTGAGTCAAGCG-1     avid_1          0            0               1
#> 1_AAACCTGCAGTTTACG-1     avid_1          4            4               0
#>                      seurat_clusters    UMAP_1     UMAP_2 clonotype_id
#> 1_AAACCTGAGTAGGTGC-1               1  2.561658 -0.6746224         <NA>
#> 1_AAACCTGAGTCAAGCG-1               1 -3.402685  6.8385974         <NA>
#> 1_AAACCTGCAGTTTACG-1               0  1.190111 -2.1084603 clonotype537
#>                                        v_gene         d_gene
#> 1_AAACCTGAGTAGGTGC-1                     <NA>           <NA>
#> 1_AAACCTGAGTCAAGCG-1                     <NA>           <NA>
#> 1_AAACCTGCAGTTTACG-1 IGHV1-9;IGKV5-43;IGKV3-2 None;None;None
#>                                 j_gene         c_gene      chains isotype
#> 1_AAACCTGAGTAGGTGC-1              <NA>           <NA>        <NA>    <NA>
#> 1_AAACCTGAGTCAAGCG-1              <NA>           <NA>        <NA>    <NA>
#> 1_AAACCTGCAGTTTACG-1 IGHJ4;IGKJ2;IGKJ1 IGHM;IGKC;IGKC IGH;IGK;IGK    IGHM
#>                                                      cdr3
#> 1_AAACCTGAGTAGGTGC-1                                 <NA>
#> 1_AAACCTGAGTCAAGCG-1                                 <NA>
#> 1_AAACCTGCAGTTTACG-1 CATLWYDAMDYW;CQQSNSWPYTF;CQQSKEVPRTF
#>                                                                                                                       cdr3_nt
#> 1_AAACCTGAGTAGGTGC-1                                                                                                     <NA>
#> 1_AAACCTGAGTCAAGCG-1                                                                                                     <NA>
#> 1_AAACCTGCAGTTTACG-1 TGTGCAACTCTATGGTACGATGCTATGGACTACTGG;TGTCAACAGAGTAACAGCTGGCCGTACACGTTC;TGTCAGCAAAGTAAGGAGGTTCCTCGGACGTTC
#>                            reads    umis     productive    full_length
#> 1_AAACCTGAGTAGGTGC-1        <NA>    <NA>           <NA>           <NA>
#> 1_AAACCTGAGTCAAGCG-1        <NA>    <NA>           <NA>           <NA>
#> 1_AAACCTGCAGTTTACG-1 144;460;156 7;28;10 TRUE;TRUE;TRUE TRUE;TRUE;TRUE
#>                      paired n_insertion n_deletion n_mismatch cdr3_length
#> 1_AAACCTGAGTAGGTGC-1     NA        <NA>       <NA>       <NA>        <NA>
#> 1_AAACCTGAGTCAAGCG-1     NA        <NA>       <NA>       <NA>        <NA>
#> 1_AAACCTGCAGTTTACG-1   TRUE       4;0;0      0;4;3      4;2;0    12;11;11
#>                      cdr3_nt_length n_chains
#> 1_AAACCTGAGTAGGTGC-1           <NA>       NA
#> 1_AAACCTGAGTCAAGCG-1           <NA>       NA
#> 1_AAACCTGCAGTTTACG-1       36;33;33        3

# Only include V(D)J data for cells that have an IGH, IGK, and IGL chain
res <- filter_vdj(
  vdj_sce,
  all(c("IGH", "IGK", "IGL") %in% chains)
)

head(res@colData, 1)
#> DataFrame with 1 row and 27 columns
#>                       orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#>                      <character>  <numeric>    <integer>        <factor>
#> 1_AAACCTGAGTAGGTGC-1      avid_1          1            1               1
#>                      seurat_clusters    UMAP_1    UMAP_2 clonotype_id
#>                             <factor> <numeric> <numeric>  <character>
#> 1_AAACCTGAGTAGGTGC-1               1   2.56166 -0.674622           NA
#>                           v_gene      d_gene      j_gene      c_gene
#>                      <character> <character> <character> <character>
#> 1_AAACCTGAGTAGGTGC-1          NA          NA          NA          NA
#>                           chains     isotype        cdr3     cdr3_nt
#>                      <character> <character> <character> <character>
#> 1_AAACCTGAGTAGGTGC-1          NA          NA          NA          NA
#>                            reads        umis  productive full_length
#>                      <character> <character> <character> <character>
#> 1_AAACCTGAGTAGGTGC-1          NA          NA          NA          NA
#>                         paired n_insertion  n_deletion  n_mismatch
#>                      <logical> <character> <character> <character>
#> 1_AAACCTGAGTAGGTGC-1        NA          NA          NA          NA
#>                      cdr3_length cdr3_nt_length  n_chains
#>                      <character>    <character> <integer>
#> 1_AAACCTGAGTAGGTGC-1          NA             NA        NA

# Only include V(D)J data for heavy chains
res <- filter_vdj(vdj_so, chains == "IGH")

head(res@meta.data, 3)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAACCTGAGTAGGTGC-1     avid_1          1            1               1
#> 1_AAACCTGAGTCAAGCG-1     avid_1          0            0               1
#> 1_AAACCTGCAGTTTACG-1     avid_1          4            4               0
#>                      seurat_clusters    UMAP_1     UMAP_2 clonotype_id
#> 1_AAACCTGAGTAGGTGC-1               1  2.561658 -0.6746224         <NA>
#> 1_AAACCTGAGTCAAGCG-1               1 -3.402685  6.8385974         <NA>
#> 1_AAACCTGCAGTTTACG-1               0  1.190111 -2.1084603 clonotype537
#>                       v_gene d_gene j_gene c_gene chains isotype
#> 1_AAACCTGAGTAGGTGC-1    <NA>   <NA>   <NA>   <NA>   <NA>    <NA>
#> 1_AAACCTGAGTCAAGCG-1    <NA>   <NA>   <NA>   <NA>   <NA>    <NA>
#> 1_AAACCTGCAGTTTACG-1 IGHV1-9   None  IGHJ4   IGHM    IGH    IGHM
#>                              cdr3                              cdr3_nt
#> 1_AAACCTGAGTAGGTGC-1         <NA>                                 <NA>
#> 1_AAACCTGAGTCAAGCG-1         <NA>                                 <NA>
#> 1_AAACCTGCAGTTTACG-1 CATLWYDAMDYW TGTGCAACTCTATGGTACGATGCTATGGACTACTGG
#>                      reads umis productive full_length paired n_insertion
#> 1_AAACCTGAGTAGGTGC-1  <NA> <NA>       <NA>        <NA>     NA        <NA>
#> 1_AAACCTGAGTCAAGCG-1  <NA> <NA>       <NA>        <NA>     NA        <NA>
#> 1_AAACCTGCAGTTTACG-1   144    7       TRUE        TRUE   TRUE           4
#>                      n_deletion n_mismatch cdr3_length cdr3_nt_length
#> 1_AAACCTGAGTAGGTGC-1       <NA>       <NA>        <NA>           <NA>
#> 1_AAACCTGAGTCAAGCG-1       <NA>       <NA>        <NA>           <NA>
#> 1_AAACCTGCAGTTTACG-1          0          4          12             36
#>                      n_chains
#> 1_AAACCTGAGTAGGTGC-1       NA
#> 1_AAACCTGAGTCAAGCG-1       NA
#> 1_AAACCTGCAGTTTACG-1        3

# Remove chains that only have 1 UMI for support
res <- filter_vdj(vdj_sce, umis > 1)

head(res@colData, 1)
#> DataFrame with 1 row and 27 columns
#>                       orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#>                      <character>  <numeric>    <integer>        <factor>
#> 1_AAACCTGAGTAGGTGC-1      avid_1          1            1               1
#>                      seurat_clusters    UMAP_1    UMAP_2 clonotype_id
#>                             <factor> <numeric> <numeric>  <character>
#> 1_AAACCTGAGTAGGTGC-1               1   2.56166 -0.674622   clonotype1
#>                           v_gene      d_gene      j_gene      c_gene
#>                      <character> <character> <character> <character>
#> 1_AAACCTGAGTAGGTGC-1    IGKV5-43        None       IGKJ2        IGKC
#>                           chains     isotype        cdr3
#>                      <character> <character> <character>
#> 1_AAACCTGAGTAGGTGC-1         IGK        None CQQSNSWPYTF
#>                                     cdr3_nt       reads        umis
#>                                 <character> <character> <character>
#> 1_AAACCTGAGTAGGTGC-1 TGTCAACAGAGTAACAGCTG..        1054          70
#>                       productive full_length    paired n_insertion
#>                      <character> <character> <logical> <character>
#> 1_AAACCTGAGTAGGTGC-1        TRUE        TRUE     FALSE           0
#>                       n_deletion  n_mismatch cdr3_length cdr3_nt_length
#>                      <character> <character> <character>    <character>
#> 1_AAACCTGAGTAGGTGC-1           4           2          11             33
#>                       n_chains
#>                      <integer>
#> 1_AAACCTGAGTAGGTGC-1         1

# Filter based on cell barcode
res <- filter_vdj(vdj_so, .cell_id == "1_AAACCTGAGTAGGTGC-1")

head(res@meta.data, 3)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAACCTGAGTAGGTGC-1     avid_1          1            1               1
#> 1_AAACCTGAGTCAAGCG-1     avid_1          0            0               1
#> 1_AAACCTGCAGTTTACG-1     avid_1          4            4               0
#>                      seurat_clusters    UMAP_1     UMAP_2 clonotype_id
#> 1_AAACCTGAGTAGGTGC-1               1  2.561658 -0.6746224   clonotype1
#> 1_AAACCTGAGTCAAGCG-1               1 -3.402685  6.8385974         <NA>
#> 1_AAACCTGCAGTTTACG-1               0  1.190111 -2.1084603         <NA>
#>                        v_gene d_gene j_gene c_gene chains isotype
#> 1_AAACCTGAGTAGGTGC-1 IGKV5-43   None  IGKJ2   IGKC    IGK    None
#> 1_AAACCTGAGTCAAGCG-1     <NA>   <NA>   <NA>   <NA>   <NA>    <NA>
#> 1_AAACCTGCAGTTTACG-1     <NA>   <NA>   <NA>   <NA>   <NA>    <NA>
#>                             cdr3                           cdr3_nt reads
#> 1_AAACCTGAGTAGGTGC-1 CQQSNSWPYTF TGTCAACAGAGTAACAGCTGGCCGTACACGTTC  1054
#> 1_AAACCTGAGTCAAGCG-1        <NA>                              <NA>  <NA>
#> 1_AAACCTGCAGTTTACG-1        <NA>                              <NA>  <NA>
#>                      umis productive full_length paired n_insertion
#> 1_AAACCTGAGTAGGTGC-1   70       TRUE        TRUE  FALSE           0
#> 1_AAACCTGAGTCAAGCG-1 <NA>       <NA>        <NA>     NA        <NA>
#> 1_AAACCTGCAGTTTACG-1 <NA>       <NA>        <NA>     NA        <NA>
#>                      n_deletion n_mismatch cdr3_length cdr3_nt_length
#> 1_AAACCTGAGTAGGTGC-1          4          2          11             33
#> 1_AAACCTGAGTCAAGCG-1       <NA>       <NA>        <NA>           <NA>
#> 1_AAACCTGCAGTTTACG-1       <NA>       <NA>        <NA>           <NA>
#>                      n_chains
#> 1_AAACCTGAGTAGGTGC-1        1
#> 1_AAACCTGAGTCAAGCG-1       NA
#> 1_AAACCTGCAGTTTACG-1       NA