Import V(D)J data

Import V(D)J data

Usage

import_vdj(
  input = NULL,
  vdj_dir,
  prefix = "",
  cell_prefix = NULL,
  filter_chains = TRUE,
  filter_paired = FALSE,
  define_clonotypes = NULL,
  include_indels = TRUE,
  sep = ";"
)

Arguments

input

Object containing single cell data, if set to NULL a data.frame containing V(D)J results will be returned

vdj_dir

Directory containing the output from cellranger vdj. A vector or named vector can be given to load data from multiple runs. If a named vector is given, the cell barcodes will be prefixed with the provided names. This mimics the behavior of Seurat::Read10X(). Cell barcode prefixes can also be provided using the cell_prefix argument.

prefix

Prefix to add to new columns

cell_prefix

Prefix to add to cell barcodes, this is helpful when loading data from multiple runs into a single object. If NULL, cell barcode prefixes are automatically generated in a similar manner as Seurat::Read10X().

For the V(D)J data to be successfully added to the object, the cell prefixes must match the prefixes that are already present in the object. If the cell prefixes are incorrect, import_vdj will be unable to assign the V(D)J data to the correct cells.

filter_chains

Only include chains with at least one productive and full length contig.

filter_paired

Only include clonotypes with paired chains. For TCR data each clonotype must have at least one TRA and TRB chain, for BCR data each clonotype must have at least one IGH chain and at least one IGK or IGL chain.

define_clonotypes

Define clonotype IDs based on V(D)J data. This is useful if the V(D)J datasets being loaded do not have consistent clonotype IDs, i.e., clonotype1 is not the same across samples. Possible values are:

cdr3aa, define clonotypes based on the CDR3 amino acid sequence

cdr3nt, define clonotypes based on the CDR3 nucleotide sequence

cdr3_gene, define clonotypes based on the combination of the CDR3 nucleotide sequence and the V(D)J genes.

When defining clonotypes, only productive full length chains will be used. Set to NULL (default) to use the clonotype IDs already present in the input data.

include_indels

Include the number of insertions/deletions for each chain. This requires the concat_ref.bam file from cellranger vdj to be present the directory provided to vdj_dir. If include_indels is TRUE, filter_chains is also automatically set TRUE since indel data is only available for productive chains.

sep

Separator to use for storing per cell V(D)J data

Value

Single cell object or data.frame with added V(D)J data

Examples

# Loading multiple datasets
vdj_dir <- c(
  system.file("extdata/bcr_1", package = "djvdj"),
  system.file("extdata/bcr_2", package = "djvdj")
)

vdj_so <- import_vdj(tiny_so, vdj_dir)

head(vdj_so@meta.data, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAACCTGAGTAGGTGC-1     avid_1          1            1               1
#>                      seurat_clusters   UMAP_1     UMAP_2 clonotype_id   v_gene
#> 1_AAACCTGAGTAGGTGC-1               1 2.561658 -0.6746224   clonotype1 IGKV5-43
#>                      d_gene j_gene c_gene chains isotype        cdr3
#> 1_AAACCTGAGTAGGTGC-1   None  IGKJ2   IGKC    IGK    None CQQSNSWPYTF
#>                                                cdr3_nt reads umis productive
#> 1_AAACCTGAGTAGGTGC-1 TGTCAACAGAGTAACAGCTGGCCGTACACGTTC  1054   70       TRUE
#>                      full_length paired n_insertion n_deletion n_mismatch
#> 1_AAACCTGAGTAGGTGC-1        TRUE  FALSE           0          4          2
#>                      cdr3_length cdr3_nt_length n_chains
#> 1_AAACCTGAGTAGGTGC-1          11             33        1

# Specifying cell prefixes
# if cell prefixes are not specified when loading multiple datasets,
# prefixes will be automatically generated in a similar manner as
# Seurat::Read10X
vdj_so <- import_vdj(
  tiny_so,
  vdj_dir = vdj_dir,
  cell_prefix = c("1_", "2_")
)

head(vdj_so@meta.data, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAACCTGAGTAGGTGC-1     avid_1          1            1               1
#>                      seurat_clusters   UMAP_1     UMAP_2 clonotype_id   v_gene
#> 1_AAACCTGAGTAGGTGC-1               1 2.561658 -0.6746224   clonotype1 IGKV5-43
#>                      d_gene j_gene c_gene chains isotype        cdr3
#> 1_AAACCTGAGTAGGTGC-1   None  IGKJ2   IGKC    IGK    None CQQSNSWPYTF
#>                                                cdr3_nt reads umis productive
#> 1_AAACCTGAGTAGGTGC-1 TGTCAACAGAGTAACAGCTGGCCGTACACGTTC  1054   70       TRUE
#>                      full_length paired n_insertion n_deletion n_mismatch
#> 1_AAACCTGAGTAGGTGC-1        TRUE  FALSE           0          4          2
#>                      cdr3_length cdr3_nt_length n_chains
#> 1_AAACCTGAGTAGGTGC-1          11             33        1

# Specifying cell prefixes using vector names
# if a named vector is passed, the names will be used as the cell prefixes
vdj_dir <- c(
  "1_" = system.file("extdata/bcr_1", package = "djvdj"),
  "2_" = system.file("extdata/bcr_2", package = "djvdj")
)

vdj_so <- import_vdj(tiny_so, vdj_dir)

head(vdj_so@meta.data, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAACCTGAGTAGGTGC-1     avid_1          1            1               1
#>                      seurat_clusters   UMAP_1     UMAP_2 clonotype_id   v_gene
#> 1_AAACCTGAGTAGGTGC-1               1 2.561658 -0.6746224   clonotype1 IGKV5-43
#>                      d_gene j_gene c_gene chains isotype        cdr3
#> 1_AAACCTGAGTAGGTGC-1   None  IGKJ2   IGKC    IGK    None CQQSNSWPYTF
#>                                                cdr3_nt reads umis productive
#> 1_AAACCTGAGTAGGTGC-1 TGTCAACAGAGTAACAGCTGGCCGTACACGTTC  1054   70       TRUE
#>                      full_length paired n_insertion n_deletion n_mismatch
#> 1_AAACCTGAGTAGGTGC-1        TRUE  FALSE           0          4          2
#>                      cdr3_length cdr3_nt_length n_chains
#> 1_AAACCTGAGTAGGTGC-1          11             33        1

# Only include V(D)J data for productive full length chains
vdj_so <- import_vdj(
  tiny_so,
  vdj_dir = vdj_dir,
  filter_chains = TRUE
)

head(vdj_so@meta.data, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAACCTGAGTAGGTGC-1     avid_1          1            1               1
#>                      seurat_clusters   UMAP_1     UMAP_2 clonotype_id   v_gene
#> 1_AAACCTGAGTAGGTGC-1               1 2.561658 -0.6746224   clonotype1 IGKV5-43
#>                      d_gene j_gene c_gene chains isotype        cdr3
#> 1_AAACCTGAGTAGGTGC-1   None  IGKJ2   IGKC    IGK    None CQQSNSWPYTF
#>                                                cdr3_nt reads umis productive
#> 1_AAACCTGAGTAGGTGC-1 TGTCAACAGAGTAACAGCTGGCCGTACACGTTC  1054   70       TRUE
#>                      full_length paired n_insertion n_deletion n_mismatch
#> 1_AAACCTGAGTAGGTGC-1        TRUE  FALSE           0          4          2
#>                      cdr3_length cdr3_nt_length n_chains
#> 1_AAACCTGAGTAGGTGC-1          11             33        1

# Only include V(D)J data for cells with paired chains
vdj_so <- import_vdj(
  tiny_so,
  vdj_dir = vdj_dir,
  filter_paired = TRUE
)

head(vdj_so@meta.data, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAACCTGAGTAGGTGC-1     avid_1          1            1               1
#>                      seurat_clusters   UMAP_1     UMAP_2 clonotype_id v_gene
#> 1_AAACCTGAGTAGGTGC-1               1 2.561658 -0.6746224         <NA>   <NA>
#>                      d_gene j_gene c_gene chains isotype cdr3 cdr3_nt reads
#> 1_AAACCTGAGTAGGTGC-1   <NA>   <NA>   <NA>   <NA>    <NA> <NA>    <NA>  <NA>
#>                      umis productive full_length paired n_insertion n_deletion
#> 1_AAACCTGAGTAGGTGC-1 <NA>       <NA>        <NA>     NA        <NA>       <NA>
#>                      n_mismatch cdr3_length cdr3_nt_length n_chains
#> 1_AAACCTGAGTAGGTGC-1       <NA>        <NA>           <NA>       NA

# Defining clonotypes
# this is useful if the original clonotype IDs are not consistent across
# datasets, i.e. clonotype1 is not the same for all samples
vdj_so <- import_vdj(
  tiny_so,
  vdj_dir = vdj_dir,
  define_clonotypes = "cdr3_gene"
)

head(vdj_so@meta.data, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAACCTGAGTAGGTGC-1     avid_1          1            1               1
#>                      seurat_clusters   UMAP_1     UMAP_2 clonotype_id   v_gene
#> 1_AAACCTGAGTAGGTGC-1               1 2.561658 -0.6746224  clonotype16 IGKV5-43
#>                      d_gene j_gene c_gene chains isotype        cdr3
#> 1_AAACCTGAGTAGGTGC-1   None  IGKJ2   IGKC    IGK    None CQQSNSWPYTF
#>                                                cdr3_nt reads umis productive
#> 1_AAACCTGAGTAGGTGC-1 TGTCAACAGAGTAACAGCTGGCCGTACACGTTC  1054   70       TRUE
#>                      full_length paired n_insertion n_deletion n_mismatch
#> 1_AAACCTGAGTAGGTGC-1        TRUE  FALSE           0          4          2
#>                      cdr3_length cdr3_nt_length n_chains
#> 1_AAACCTGAGTAGGTGC-1          11             33        1

# Omit indel information for each chain
# this information will be included if the file concat_ref.bam is present
# to speed up data import, omit indel information
vdj_so <- import_vdj(
  tiny_so,
  vdj_dir = vdj_dir,
  include_indels = FALSE
)

head(vdj_so@meta.data, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAACCTGAGTAGGTGC-1     avid_1          1            1               1
#>                      seurat_clusters   UMAP_1     UMAP_2 clonotype_id   v_gene
#> 1_AAACCTGAGTAGGTGC-1               1 2.561658 -0.6746224   clonotype1 IGKV5-43
#>                      d_gene j_gene c_gene chains isotype        cdr3
#> 1_AAACCTGAGTAGGTGC-1   None  IGKJ2   IGKC    IGK    None CQQSNSWPYTF
#>                                                cdr3_nt reads umis productive
#> 1_AAACCTGAGTAGGTGC-1 TGTCAACAGAGTAACAGCTGGCCGTACACGTTC  1054   70       TRUE
#>                      full_length paired cdr3_length cdr3_nt_length n_chains
#> 1_AAACCTGAGTAGGTGC-1        TRUE  FALSE          11             33        1

# Loading both BCR and TCR data
# load each dataset separately and assign unique column names using the
# prefix argument
bcr_dir <- system.file("extdata/bcr_1", package = "djvdj")
tcr_dir <- system.file("extdata/tcr_1", package = "djvdj")

vdj_so <- import_vdj(
  tiny_so,
  vdj_dir     = bcr_dir,
  prefix      = "bcr_",
  cell_prefix = "1_"
)

vdj_so <- import_vdj(
  vdj_so,
  vdj_dir     = tcr_dir,
  prefix      = "tcr_",
  cell_prefix = "1_"
)

head(vdj_so@meta.data, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAACCTGAGTAGGTGC-1     avid_1          1            1               1
#>                      seurat_clusters   UMAP_1     UMAP_2 bcr_clonotype_id
#> 1_AAACCTGAGTAGGTGC-1               1 2.561658 -0.6746224       clonotype1
#>                      bcr_v_gene bcr_d_gene bcr_j_gene bcr_c_gene bcr_chains
#> 1_AAACCTGAGTAGGTGC-1   IGKV5-43       None      IGKJ2       IGKC        IGK
#>                      bcr_isotype    bcr_cdr3                       bcr_cdr3_nt
#> 1_AAACCTGAGTAGGTGC-1        None CQQSNSWPYTF TGTCAACAGAGTAACAGCTGGCCGTACACGTTC
#>                      bcr_reads bcr_umis bcr_productive bcr_full_length
#> 1_AAACCTGAGTAGGTGC-1      1054       70           TRUE            TRUE
#>                      bcr_paired bcr_n_insertion bcr_n_deletion bcr_n_mismatch
#> 1_AAACCTGAGTAGGTGC-1      FALSE               0              4              2
#>                      bcr_cdr3_length bcr_cdr3_nt_length bcr_n_chains
#> 1_AAACCTGAGTAGGTGC-1              11                 33            1
#>                      tcr_clonotype_id       tcr_v_gene tcr_d_gene
#> 1_AAACCTGAGTAGGTGC-1     clonotype116 TRAV12D-2;TRBV20  None;None
#>                          tcr_j_gene tcr_c_gene tcr_chains
#> 1_AAACCTGAGTAGGTGC-1 TRAJ40;TRBJ2-5  TRAC;None    TRA;TRB
#>                                          tcr_cdr3
#> 1_AAACCTGAGTAGGTGC-1 CALSDRWDSGNYKYVF;CGAGAGDTQYF
#>                                                                                             tcr_cdr3_nt
#> 1_AAACCTGAGTAGGTGC-1 TGTGCTTTGAGTGATCGATGGGACTCAGGAAACTACAAATACGTCTTT;TGTGGTGCTGGGGCGGGGGACACCCAGTACTTT
#>                      tcr_reads tcr_umis tcr_productive tcr_full_length
#> 1_AAACCTGAGTAGGTGC-1   244;856      3;6      TRUE;TRUE       TRUE;TRUE
#>                      tcr_paired tcr_n_insertion tcr_n_deletion tcr_n_mismatch
#> 1_AAACCTGAGTAGGTGC-1       TRUE             4;0            0;2            5;6
#>                      tcr_cdr3_length tcr_cdr3_nt_length tcr_n_chains
#> 1_AAACCTGAGTAGGTGC-1           16;11              48;33            2

# Using import_vdj outside of Seurat
# SingleCellExperiment objects are also compatible, or if an input object is
# omitted, a data.frame containing the V(D)J data will be returned
vdj_sce <- import_vdj(tiny_sce, vdj_dir)

head(vdj_sce@colData, 1)
#> DataFrame with 1 row and 27 columns
#>                       orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#>                      <character>  <numeric>    <integer>        <factor>
#> 1_AAACCTGAGTAGGTGC-1      avid_1          1            1               1
#>                      seurat_clusters    UMAP_1    UMAP_2 clonotype_id
#>                             <factor> <numeric> <numeric>  <character>
#> 1_AAACCTGAGTAGGTGC-1               1   2.56166 -0.674622   clonotype1
#>                           v_gene      d_gene      j_gene      c_gene
#>                      <character> <character> <character> <character>
#> 1_AAACCTGAGTAGGTGC-1    IGKV5-43        None       IGKJ2        IGKC
#>                           chains     isotype        cdr3                cdr3_nt
#>                      <character> <character> <character>            <character>
#> 1_AAACCTGAGTAGGTGC-1         IGK        None CQQSNSWPYTF TGTCAACAGAGTAACAGCTG..
#>                            reads        umis  productive full_length    paired
#>                      <character> <character> <character> <character> <logical>
#> 1_AAACCTGAGTAGGTGC-1        1054          70        TRUE        TRUE     FALSE
#>                      n_insertion  n_deletion  n_mismatch cdr3_length
#>                      <character> <character> <character> <character>
#> 1_AAACCTGAGTAGGTGC-1           0           4           2          11
#>                      cdr3_nt_length  n_chains
#>                         <character> <integer>
#> 1_AAACCTGAGTAGGTGC-1             33         1

vdj_df <- import_vdj(vdj_dir = vdj_dir)

head(vdj_df, 1)
#>                       clonotype_id                     v_gene         d_gene
#> 1_CTTCTCTTCACCTTAT-1 clonotype2350 IGHV5-17;IGKV5-43;IGKV5-43 None;None;None
#>                                 j_gene         c_gene      chains isotype
#> 1_CTTCTCTTCACCTTAT-1 IGHJ4;IGKJ2;IGKJ2 IGHA;IGKC;IGKC IGH;IGK;IGK    IGHA
#>                                                             cdr3
#> 1_CTTCTCTTCACCTTAT-1 CARPYYYTTIYVDYAMAYW;CQQSNSWPYTF;CQQSNCWPFTF
#>                                                                                                                                            cdr3_nt
#> 1_CTTCTCTTCACCTTAT-1 TGTGCAAGGCCTTATTACTACACTACTATCTACGTTGACTACGCTATGGCCTACTGG;TGTCAACAGAGTAACAGCTGGCCGTACACGTTC;TGTCAACAGAGTAACTGCTGGCCGTTCACGTTC
#>                                 reads          umis     productive
#> 1_CTTCTCTTCACCTTAT-1 14930;18030;4334 1112;1408;351 TRUE;TRUE;TRUE
#>                         full_length paired n_insertion n_deletion n_mismatch
#> 1_CTTCTCTTCACCTTAT-1 TRUE;TRUE;TRUE   TRUE      25;0;0      0;4;4     14;2;6
#>                      cdr3_length cdr3_nt_length n_chains
#> 1_CTTCTCTTCACCTTAT-1    19;11;11       57;33;33        3