Modify object meta.data

Modify object meta.data

mutate_meta(input, fn, ...)

Arguments

input

Single cell object or data.frame containing V(D)J data. If a data.frame is provided, the cell barcodes should be stored as row names.

fn

Function to use for modifying object meta.data. This can be either a function, e.g. mean, or a purrr-style lambda, e.g. ~ mean(.x, na.rm = TRUE) where ".x" refers to the meta.data table.

...

Additional arguments to pass to the provided function

Value

Object with mutated meta.data

Examples

# Sum two meta.data columns
# all additional arguments provided to mutate_meta() are passed directly to
# the function (in this case, dplyr::mutate())
mutate_meta(
  tiny_so,
  dplyr::mutate,
  NEW = nCount_RNA + nFeature_RNA
)
#> An object of class Seurat 
#> 200 features across 204 samples within 1 assay 
#> Active assay: RNA (200 features, 67 variable features)
#>  2 dimensional reductions calculated: pca, umap

# Pass a purrr-style lambda
# this produces the same result as the previous example
mutate_meta(
  tiny_so,
  ~ dplyr::mutate(.x, NEW = nCount_RNA + nFeature_RNA)
)
#> An object of class Seurat 
#> 200 features across 204 samples within 1 assay 
#> Active assay: RNA (200 features, 67 variable features)
#>  2 dimensional reductions calculated: pca, umap

# Modify multiple meta.data columns
mutate_meta(
  tiny_so,
  dplyr::mutate,
  NEW_1 = nCount_RNA + nFeature_RNA,
  NEW_2 = stringr::str_c(orig.ident, seurat_clusters)
)
#> An object of class Seurat 
#> 200 features across 204 samples within 1 assay 
#> Active assay: RNA (200 features, 67 variable features)
#>  2 dimensional reductions calculated: pca, umap

# Remove meta.data columns
# any function can be passed to mutate_meta(), in this example
# dplyr::select() is used to remove columns
mutate_meta(
  tiny_sce,
  dplyr::select,
  -UMAP_1
)
#> class: SingleCellExperiment 
#> dim: 200 204 
#> metadata(0):
#> assays(1): counts
#> rownames(200): Gins2 Klf11 ... Gm29265 Pkd1l1
#> rowData names(0):
#> colnames(204): 1_AAACCTGAGTAGGTGC-1 1_AAACCTGAGTCAAGCG-1 ...
#>   2_TTTACTGGTTGGTTTG-1 2_TTTGTCATCACCCGAG-1
#> colData names(7): orig.ident nCount_RNA ... seurat_clusters UMAP_2
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):

# Perform grouped operations using dplyr
# multi-line commands can be passed using brackets, just refer to the
# meta.data with ".x"
# this calculates the mean number of features for each group in the
# orig.ident meta.data column
mutate_meta(tiny_so, ~ {
  res <- dplyr::group_by(.x, orig.ident)
  res <- dplyr::mutate(res, mean_genes = mean(nFeature_RNA))
  res
})
#> An object of class Seurat 
#> 200 features across 204 samples within 1 assay 
#> Active assay: RNA (200 features, 67 variable features)
#>  2 dimensional reductions calculated: pca, umap