Modify V(D)J data in object

Modify per-chain V(D)J data for each cell. This function offers greater flexibility than summarize_vdj(), but is less user-friendly.

Usage

mutate_vdj(
  input,
  ...,
  clonotype_col = "clonotype_id",
  vdj_cols = NULL,
  return_df = FALSE,
  sep = ";"
)

Arguments

input

Single cell object or data.frame containing V(D)J data. If a data.frame is provided, cell barcodes should be stored as row names.

...

Name-value pairs to use for creating or modifying per-chain V(D)J meta.data, e.g. mean_umis = mean(umis).

To allow for modification of per-chain V(D)J data, the data for each cell is converted into a vector, e.g. 'IGH;IGK' is equivalent to c('IGH', 'IGK'). This allows R vector operations to be performed on the per-chain values. Any operations that produce a result greater than length 1 need to be returned as a list(), e.g. new_col = umis + 1 will return a new value for each chain, to prevent an error this must be written as new_col = list(umis + 1).

clonotype_col

meta.data column containing clonotype IDs. This is used to identify columns containing V(D)J data.

vdj_cols

meta.data columns containing V(D)J data to modify. If NULL, data are automatically selected by identifying columns that have NAs in the same rows as clonotype_col.

return_df

Return results as a data.frame. If FALSE, results will be added to the input object.

sep

Separator used for storing per cell V(D)J data

Value

Object with modified meta.data

Examples

# Calculate mean reads and UMIs per cell
mutate_vdj(
  vdj_sce,
  mean_umis  = mean(umis),
  mean_reads = mean(reads)
)
#> class: SingleCellExperiment 
#> dim: 200 204 
#> metadata(0):
#> assays(1): counts
#> rownames(200): Gins2 Klf11 ... Gm29265 Pkd1l1
#> rowData names(0):
#> colnames(204): 1_AAACCTGAGTAGGTGC-1 1_AAACCTGAGTCAAGCG-1 ...
#>   2_TTTACTGGTTGGTTTG-1 2_TTTGTCATCACCCGAG-1
#> colData names(30): orig.ident nCount_RNA ... mean_umis mean_reads
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):

# Calculate the total number of insertions + deletions for each chain
# we have to wrap our expression in list() since a value is returned for
# each chain
mutate_vdj(
  vdj_sce,
  indels = list(n_insertion + n_deletion)
)
#> class: SingleCellExperiment 
#> dim: 200 204 
#> metadata(0):
#> assays(1): counts
#> rownames(200): Gins2 Klf11 ... Gm29265 Pkd1l1
#> rowData names(0):
#> colnames(204): 1_AAACCTGAGTAGGTGC-1 1_AAACCTGAGTCAAGCG-1 ...
#>   2_TTTACTGGTTGGTTTG-1 2_TTTGTCATCACCCGAG-1
#> colData names(29): orig.ident nCount_RNA ... n_chains indels
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):

# Create a new column showing the unique chains for each cell
mutate_vdj(
  vdj_sce,
  unique_chains = stringr::str_c(unique(chains), collapse = "_")
)
#> class: SingleCellExperiment 
#> dim: 200 204 
#> metadata(0):
#> assays(1): counts
#> rownames(200): Gins2 Klf11 ... Gm29265 Pkd1l1
#> rowData names(0):
#> colnames(204): 1_AAACCTGAGTAGGTGC-1 1_AAACCTGAGTCAAGCG-1 ...
#>   2_TTTACTGGTTGGTTTG-1 2_TTTGTCATCACCCGAG-1
#> colData names(29): orig.ident nCount_RNA ... n_chains unique_chains
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):

# Determine which cells have both an IGK and IGL chain
mutate_vdj(
  vdj_sce,
  both_light = all(c("IGK", "IGL") %in% chains)
)
#> class: SingleCellExperiment 
#> dim: 200 204 
#> metadata(0):
#> assays(1): counts
#> rownames(200): Gins2 Klf11 ... Gm29265 Pkd1l1
#> rowData names(0):
#> colnames(204): 1_AAACCTGAGTAGGTGC-1 1_AAACCTGAGTCAAGCG-1 ...
#>   2_TTTACTGGTTGGTTTG-1 2_TTTGTCATCACCCGAG-1
#> colData names(29): orig.ident nCount_RNA ... n_chains both_light
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):

# Determine which cells have multiple light chains
mutate_vdj(
  vdj_so,
  multi_light = sum(chains %in% c("IGK", "IGL")) > 1
)
#> An object of class Seurat 
#> 200 features across 204 samples within 1 assay 
#> Active assay: RNA (200 features, 67 variable features)
#>  2 dimensional reductions calculated: pca, umap