Create bar graphs summarizing cell counts for clusters

Create bar graphs summarizing cell counts for clusters

plot_cell_count(
  sobj_in,
  x,
  fill_col = NULL,
  facet_col = NULL,
  yaxis = "fraction",
  plot_colors = NULL,
  plot_lvls = NULL,
  na_color = "grey90",
  n_label = TRUE,
  label_aes = list(),
  facet_rows = 1,
  facet_scales = "free_x",
  ...
)

Arguments

sobj_in
Seurat object or data.frame containing data

x
meta.data column containing clusters to use for creating bars

fill_col
meta.data column to use for coloring bars

facet_col
meta.data column containing groups to use for spliting plot into facets

yaxis
Units to plot on the y-axis, either "fraction" or "counts"

plot_colors
Character vector containing colors for plotting

plot_lvls
Character vector containing levels for ordering

na_color
Color to use for missing values

n_label
Include label showing the number of cells represented by each bar

label_aes
Named list providing additional label aesthetics (color, size, etc.)

facet_rows
The number of facet rows. Use this argument if facet_col is specified

facet_scales
If facet_col is used, this argument passes a scales specification to facet_wrap. Can be "fixed", "free", "free_x", or "free_y"

...
Additional arguments to pass to ggplot

Value

ggplot object