Create MDS plot

Perform multidimensional scaling and plot results

Usage

plot_mds(
  input,
  data_col,
  cluster_col,
  method = abdiv::jaccard,
  chain = NULL,
  chain_col = "chains",
  plot_colors = NULL,
  plot_lvls = names(plot_colors),
  label_points = TRUE,
  sep = ";",
  ...
)

Arguments

input

Single cell object or data.frame containing V(D)J data. If a data.frame is provided, the cell barcodes should be stored as row names.

data_col

meta.data column containing values to use for calculating pairwise similarity between clusters, e.g. 'clonotype_id'

cluster_col

meta.data column containing cluster IDs to use for calculating overlap

method

Method to use for comparing clusters and calculating MDS coordinates, available methods include:

'jaccard', Jaccard dissimilarity index implemented with abdiv::jaccard()

'horn_morisita', Horn-Morisita index implemented with abdiv::horn_morisita()

chain

Chain to use for calculating gene usage. Set to NULL to include all chains.

chain_col

meta.data column containing chains for each cell

plot_colors

Character vector containing colors for plotting

plot_lvls

Levels to use for ordering clusters

label_points

Label points on plot

sep

Separator used for storing per-chain V(D)J data for each cell

...

Additional arguments to pass to ggplot2::geom_point()

See also