Create sequence logos for clusters

Create sequence logos for clusters

Usage

plot_motifs(
  input,
  data_col = "cdr3",
  cluster_col = NULL,
  chain,
  plot_colors = NULL,
  plot_lvls = names(plot_colors),
  chain_col = "chains",
  width = 0.75,
  align_end = "5",
  panel_nrow = NULL,
  sep = ";",
  ...
)

Arguments

input

Single cell object or data.frame containing V(D)J data. If a data.frame is provided, the cell barcodes should be stored as row names.

data_col

meta.data column containing sequences to use for plotting.

cluster_col

meta.data column containing cluster IDs to use for grouping cells.

chain

Chain to use for clustering CDR3 sequences. Cells with more than one of the provided chain will be excluded from the analysis.

plot_colors

Character vector containing colors for plotting

plot_lvls

Character vector containing levels for ordering

chain_col

meta.data column containing chains for each cell.

width

Integer specifying how many residues to include, sequences longer than width will be get trimmed based on the align_end argument, sequences shorter than width will get removed. If a fraction is provided, the width cutoff is set based on percent rank, i.e. a value of 0.75 would select a width where at least 75% of sequences are longer than the cutoff.

align_end

End to use for aligning sequences, specify '5' or '3' to align sequences at the 5' or 3' end when plotting.

panel_nrow

The number of rows to use for arranging plot panels

sep

Separator used for storing per cell V(D)J data

...

Additional parameters to pass to ggseqlogo::geom_logo()

See also