Plot V(D)J gene usage

Plot V(D)J gene usage

plot_usage(
  sobj_in,
  gene_cols,
  cluster_col = NULL,
  chain = NULL,
  type = "heatmap",
  plot_colors = NULL,
  plot_genes = NULL,
  n_genes = NULL,
  clust_lvls = NULL,
  yaxis = "percent",
  chain_col = NULL,
  sep = ";",
  ...
)

Arguments

sobj_in
Seurat object containing V(D)J data

gene_cols
meta.data column containing genes used for each clonotype

cluster_col
meta.data column containing cell clusters to use for calculating gene usage

chain
Chain to use for calculating gene usage. Set to NULL to include all chains.

type
Type of plot to create, can be either 'heatmap' or 'bar'

plot_colors
Character vector containing colors for plotting

plot_genes
Character vector containing genes to plot

n_genes
Number of top genes to plot based on average usage

clust_lvls
Levels to use for ordering clusters

yaxis
Units to plot on the y-axis, either "frequency" or "percent"

chain_col
meta.data column containing chains for each cell

sep
Separator to use for expanding gene_cols

...
Additional arguments to pass to ggplot

Value

ggplot object