Summarize V(D)J data for each cell

Summarize per-chain values for each cell using a function or purrr-style lambda. This is useful for plotting or filtering cells based on the V(D)J meta.data.

Usage

summarize_vdj(
  input,
  vdj_cols,
  fn = NULL,
  ...,
  chain = NULL,
  chain_col = "chains",
  sep = ";",
  col_names = "{.col}",
  return_df = FALSE
)

Arguments

input

Single cell object or data.frame containing V(D)J data. If a data.frame is provided, the cell barcodes should be stored as row names.

vdj_cols

meta.data column(s) containing V(D)J data to summarize for each cell

fn

Function to apply to each selected column, possible values can be either a function, e.g. mean, or a purrr-style lambda, e.g. ~ mean(.x, na.rm = TRUE). If NULL, the mean will be calculated for numeric values, non-numeric columns will be combined into a single string.

...

Additional arguments to pass to fn

chain

Chain to use for summarizing V(D)J data

chain_col

meta.data column(s) containing chains for each cell

sep

Separator used for storing per cell V(D)J data

col_names

A glue specification that describes how to name the output columns, use {.col} to refer to the original column name. If col_names is NULL, the original column names will be used.

return_df

Return results as a data.frame. If FALSE, results will be added to the input object.

Value

Object containing V(D)J data summarized for each cell

Examples

# Summarize numeric columns
# by default the mean will be calculated for numeric columns
res <- summarize_vdj(
  vdj_so,
  vdj_cols = c("n_deletion", "n_insertion")
)

head(res@meta.data, 3)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAACCTGAGTAGGTGC-1     avid_1          1            1               1
#> 1_AAACCTGAGTCAAGCG-1     avid_1          0            0               1
#> 1_AAACCTGCAGTTTACG-1     avid_1          4            4               0
#>                      seurat_clusters    UMAP_1     UMAP_2 clonotype_id
#> 1_AAACCTGAGTAGGTGC-1               1  2.561658 -0.6746224   clonotype1
#> 1_AAACCTGAGTCAAGCG-1               1 -3.402685  6.8385974 clonotype536
#> 1_AAACCTGCAGTTTACG-1               0  1.190111 -2.1084603 clonotype537
#>                                        v_gene         d_gene            j_gene
#> 1_AAACCTGAGTAGGTGC-1                 IGKV5-43           None             IGKJ2
#> 1_AAACCTGAGTCAAGCG-1      IGKV6-15;IGKV14-126      None;None       IGKJ2;IGKJ2
#> 1_AAACCTGCAGTTTACG-1 IGHV1-9;IGKV5-43;IGKV3-2 None;None;None IGHJ4;IGKJ2;IGKJ1
#>                              c_gene      chains isotype
#> 1_AAACCTGAGTAGGTGC-1           IGKC         IGK    None
#> 1_AAACCTGAGTCAAGCG-1      IGKC;IGKC     IGK;IGK    None
#> 1_AAACCTGCAGTTTACG-1 IGHM;IGKC;IGKC IGH;IGK;IGK    IGHM
#>                                                      cdr3
#> 1_AAACCTGAGTAGGTGC-1                          CQQSNSWPYTF
#> 1_AAACCTGAGTCAAGCG-1              CQQYNSYPYTF;CLQHGESPYTF
#> 1_AAACCTGCAGTTTACG-1 CATLWYDAMDYW;CQQSNSWPYTF;CQQSKEVPRTF
#>                                                                                                                       cdr3_nt
#> 1_AAACCTGAGTAGGTGC-1                                                                        TGTCAACAGAGTAACAGCTGGCCGTACACGTTC
#> 1_AAACCTGAGTCAAGCG-1                                      TGTCAGCAATATAACAGCTATCCGTACACGTTC;TGTCTACAGCATGGTGAGAGCCCGTACACGTTC
#> 1_AAACCTGCAGTTTACG-1 TGTGCAACTCTATGGTACGATGCTATGGACTACTGG;TGTCAACAGAGTAACAGCTGGCCGTACACGTTC;TGTCAGCAAAGTAAGGAGGTTCCTCGGACGTTC
#>                            reads    umis     productive    full_length paired
#> 1_AAACCTGAGTAGGTGC-1        1054      70           TRUE           TRUE  FALSE
#> 1_AAACCTGAGTCAAGCG-1     570;234   31;17      TRUE;TRUE      TRUE;TRUE  FALSE
#> 1_AAACCTGCAGTTTACG-1 144;460;156 7;28;10 TRUE;TRUE;TRUE TRUE;TRUE;TRUE   TRUE
#>                      n_insertion n_deletion n_mismatch cdr3_length
#> 1_AAACCTGAGTAGGTGC-1    0.000000   4.000000          2          11
#> 1_AAACCTGAGTCAAGCG-1    0.000000   4.000000        0;0       11;11
#> 1_AAACCTGCAGTTTACG-1    1.333333   2.333333      4;2;0    12;11;11
#>                      cdr3_nt_length n_chains
#> 1_AAACCTGAGTAGGTGC-1             33        1
#> 1_AAACCTGAGTCAAGCG-1          33;33        2
#> 1_AAACCTGCAGTTTACG-1       36;33;33        3

# Specifying a different summary function
# this calculates the median number of insertions and deletions for each
# cell
res <- summarize_vdj(
  vdj_sce,
  vdj_cols = c("n_deletion", "n_insertion"),
  fn = stats::median
)

head(res@colData, 3)
#> DataFrame with 3 rows and 27 columns
#>                       orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#>                      <character>  <numeric>    <integer>        <factor>
#> 1_AAACCTGAGTAGGTGC-1      avid_1          1            1               1
#> 1_AAACCTGAGTCAAGCG-1      avid_1          0            0               1
#> 1_AAACCTGCAGTTTACG-1      avid_1          4            4               0
#>                      seurat_clusters    UMAP_1    UMAP_2 clonotype_id
#>                             <factor> <numeric> <numeric>  <character>
#> 1_AAACCTGAGTAGGTGC-1               1   2.56166 -0.674622   clonotype1
#> 1_AAACCTGAGTCAAGCG-1               1  -3.40268  6.838597 clonotype536
#> 1_AAACCTGCAGTTTACG-1               0   1.19011 -2.108460 clonotype537
#>                                      v_gene         d_gene            j_gene
#>                                 <character>    <character>       <character>
#> 1_AAACCTGAGTAGGTGC-1               IGKV5-43           None             IGKJ2
#> 1_AAACCTGAGTCAAGCG-1    IGKV6-15;IGKV14-126      None;None       IGKJ2;IGKJ2
#> 1_AAACCTGCAGTTTACG-1 IGHV1-9;IGKV5-43;IGK.. None;None;None IGHJ4;IGKJ2;IGKJ1
#>                              c_gene      chains     isotype
#>                         <character> <character> <character>
#> 1_AAACCTGAGTAGGTGC-1           IGKC         IGK        None
#> 1_AAACCTGAGTCAAGCG-1      IGKC;IGKC     IGK;IGK        None
#> 1_AAACCTGCAGTTTACG-1 IGHM;IGKC;IGKC IGH;IGK;IGK        IGHM
#>                                        cdr3                cdr3_nt       reads
#>                                 <character>            <character> <character>
#> 1_AAACCTGAGTAGGTGC-1            CQQSNSWPYTF TGTCAACAGAGTAACAGCTG..        1054
#> 1_AAACCTGAGTCAAGCG-1 CQQYNSYPYTF;CLQHGESP.. TGTCAGCAATATAACAGCTA..     570;234
#> 1_AAACCTGCAGTTTACG-1 CATLWYDAMDYW;CQQSNSW.. TGTGCAACTCTATGGTACGA.. 144;460;156
#>                             umis     productive    full_length    paired
#>                      <character>    <character>    <character> <logical>
#> 1_AAACCTGAGTAGGTGC-1          70           TRUE           TRUE     FALSE
#> 1_AAACCTGAGTCAAGCG-1       31;17      TRUE;TRUE      TRUE;TRUE     FALSE
#> 1_AAACCTGCAGTTTACG-1     7;28;10 TRUE;TRUE;TRUE TRUE;TRUE;TRUE      TRUE
#>                      n_insertion n_deletion  n_mismatch cdr3_length
#>                        <numeric>  <numeric> <character> <character>
#> 1_AAACCTGAGTAGGTGC-1           0          4           2          11
#> 1_AAACCTGAGTCAAGCG-1           0          4         0;0       11;11
#> 1_AAACCTGCAGTTTACG-1           0          3       4;2;0    12;11;11
#>                      cdr3_nt_length  n_chains
#>                         <character> <integer>
#> 1_AAACCTGAGTAGGTGC-1             33         1
#> 1_AAACCTGAGTCAAGCG-1          33;33         2
#> 1_AAACCTGCAGTTTACG-1       36;33;33         3

# Summarize values for a specific chain
res <- summarize_vdj(
  vdj_so,
  vdj_cols = c("n_deletion", "n_insertion"),
  chain = "IGK"
)

head(res@meta.data, 3)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAACCTGAGTAGGTGC-1     avid_1          1            1               1
#> 1_AAACCTGAGTCAAGCG-1     avid_1          0            0               1
#> 1_AAACCTGCAGTTTACG-1     avid_1          4            4               0
#>                      seurat_clusters    UMAP_1     UMAP_2 clonotype_id
#> 1_AAACCTGAGTAGGTGC-1               1  2.561658 -0.6746224   clonotype1
#> 1_AAACCTGAGTCAAGCG-1               1 -3.402685  6.8385974 clonotype536
#> 1_AAACCTGCAGTTTACG-1               0  1.190111 -2.1084603 clonotype537
#>                                        v_gene         d_gene            j_gene
#> 1_AAACCTGAGTAGGTGC-1                 IGKV5-43           None             IGKJ2
#> 1_AAACCTGAGTCAAGCG-1      IGKV6-15;IGKV14-126      None;None       IGKJ2;IGKJ2
#> 1_AAACCTGCAGTTTACG-1 IGHV1-9;IGKV5-43;IGKV3-2 None;None;None IGHJ4;IGKJ2;IGKJ1
#>                              c_gene      chains isotype
#> 1_AAACCTGAGTAGGTGC-1           IGKC         IGK    None
#> 1_AAACCTGAGTCAAGCG-1      IGKC;IGKC     IGK;IGK    None
#> 1_AAACCTGCAGTTTACG-1 IGHM;IGKC;IGKC IGH;IGK;IGK    IGHM
#>                                                      cdr3
#> 1_AAACCTGAGTAGGTGC-1                          CQQSNSWPYTF
#> 1_AAACCTGAGTCAAGCG-1              CQQYNSYPYTF;CLQHGESPYTF
#> 1_AAACCTGCAGTTTACG-1 CATLWYDAMDYW;CQQSNSWPYTF;CQQSKEVPRTF
#>                                                                                                                       cdr3_nt
#> 1_AAACCTGAGTAGGTGC-1                                                                        TGTCAACAGAGTAACAGCTGGCCGTACACGTTC
#> 1_AAACCTGAGTCAAGCG-1                                      TGTCAGCAATATAACAGCTATCCGTACACGTTC;TGTCTACAGCATGGTGAGAGCCCGTACACGTTC
#> 1_AAACCTGCAGTTTACG-1 TGTGCAACTCTATGGTACGATGCTATGGACTACTGG;TGTCAACAGAGTAACAGCTGGCCGTACACGTTC;TGTCAGCAAAGTAAGGAGGTTCCTCGGACGTTC
#>                            reads    umis     productive    full_length paired
#> 1_AAACCTGAGTAGGTGC-1        1054      70           TRUE           TRUE  FALSE
#> 1_AAACCTGAGTCAAGCG-1     570;234   31;17      TRUE;TRUE      TRUE;TRUE  FALSE
#> 1_AAACCTGCAGTTTACG-1 144;460;156 7;28;10 TRUE;TRUE;TRUE TRUE;TRUE;TRUE   TRUE
#>                      n_insertion n_deletion n_mismatch cdr3_length
#> 1_AAACCTGAGTAGGTGC-1           0        4.0          2          11
#> 1_AAACCTGAGTCAAGCG-1           0        4.0        0;0       11;11
#> 1_AAACCTGCAGTTTACG-1           0        3.5      4;2;0    12;11;11
#>                      cdr3_nt_length n_chains
#> 1_AAACCTGAGTAGGTGC-1             33        1
#> 1_AAACCTGAGTCAAGCG-1          33;33        2
#> 1_AAACCTGCAGTTTACG-1       36;33;33        3

# Specifying new names for summarized columns
# use {.col} to refer to the original column name
res <- summarize_vdj(
  vdj_sce,
  vdj_cols = c("n_deletion", "n_insertion"),
  fn = stats::median,
  col_names = "median_{.col}"
)

head(res@colData, 1)
#> DataFrame with 1 row and 29 columns
#>                       orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#>                      <character>  <numeric>    <integer>        <factor>
#> 1_AAACCTGAGTAGGTGC-1      avid_1          1            1               1
#>                      seurat_clusters    UMAP_1    UMAP_2 clonotype_id
#>                             <factor> <numeric> <numeric>  <character>
#> 1_AAACCTGAGTAGGTGC-1               1   2.56166 -0.674622   clonotype1
#>                           v_gene      d_gene      j_gene      c_gene
#>                      <character> <character> <character> <character>
#> 1_AAACCTGAGTAGGTGC-1    IGKV5-43        None       IGKJ2        IGKC
#>                           chains     isotype        cdr3                cdr3_nt
#>                      <character> <character> <character>            <character>
#> 1_AAACCTGAGTAGGTGC-1         IGK        None CQQSNSWPYTF TGTCAACAGAGTAACAGCTG..
#>                            reads        umis  productive full_length    paired
#>                      <character> <character> <character> <character> <logical>
#> 1_AAACCTGAGTAGGTGC-1        1054          70        TRUE        TRUE     FALSE
#>                      n_insertion  n_deletion  n_mismatch cdr3_length
#>                      <character> <character> <character> <character>
#> 1_AAACCTGAGTAGGTGC-1           0           4           2          11
#>                      cdr3_nt_length  n_chains median_n_deletion
#>                         <character> <integer>         <numeric>
#> 1_AAACCTGAGTAGGTGC-1             33         1                 4
#>                      median_n_insertion
#>                               <numeric>
#> 1_AAACCTGAGTAGGTGC-1                  0

# Return a data.frame instead of adding the results to the input object
res <- summarize_vdj(
  vdj_so,
  vdj_cols = c("n_deletion", "n_insertion"),
  return_df = TRUE
)

head(res, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAACCTGAGTAGGTGC-1     avid_1          1            1               1
#>                      seurat_clusters   UMAP_1     UMAP_2 clonotype_id   v_gene
#> 1_AAACCTGAGTAGGTGC-1               1 2.561658 -0.6746224   clonotype1 IGKV5-43
#>                      d_gene j_gene c_gene chains isotype        cdr3
#> 1_AAACCTGAGTAGGTGC-1   None  IGKJ2   IGKC    IGK    None CQQSNSWPYTF
#>                                                cdr3_nt reads umis productive
#> 1_AAACCTGAGTAGGTGC-1 TGTCAACAGAGTAACAGCTGGCCGTACACGTTC  1054   70       TRUE
#>                      full_length paired n_insertion n_deletion n_mismatch
#> 1_AAACCTGAGTAGGTGC-1        TRUE  FALSE           0          4          2
#>                      cdr3_length cdr3_nt_length n_chains
#> 1_AAACCTGAGTAGGTGC-1          11             33        1

# Using a lambda function to summarize values
# use '.x' to refer to values in the column
# this creates a new column showing the unique chains for each cell
res <- summarize_vdj(
  vdj_sce,
  vdj_cols = "chains",
  fn = ~ paste0(unique(.x), collapse = "_"),
  col_names = "unique_chains"
)

head(res@colData, 3)
#> DataFrame with 3 rows and 28 columns
#>                       orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#>                      <character>  <numeric>    <integer>        <factor>
#> 1_AAACCTGAGTAGGTGC-1      avid_1          1            1               1
#> 1_AAACCTGAGTCAAGCG-1      avid_1          0            0               1
#> 1_AAACCTGCAGTTTACG-1      avid_1          4            4               0
#>                      seurat_clusters    UMAP_1    UMAP_2 clonotype_id
#>                             <factor> <numeric> <numeric>  <character>
#> 1_AAACCTGAGTAGGTGC-1               1   2.56166 -0.674622   clonotype1
#> 1_AAACCTGAGTCAAGCG-1               1  -3.40268  6.838597 clonotype536
#> 1_AAACCTGCAGTTTACG-1               0   1.19011 -2.108460 clonotype537
#>                                      v_gene         d_gene            j_gene
#>                                 <character>    <character>       <character>
#> 1_AAACCTGAGTAGGTGC-1               IGKV5-43           None             IGKJ2
#> 1_AAACCTGAGTCAAGCG-1    IGKV6-15;IGKV14-126      None;None       IGKJ2;IGKJ2
#> 1_AAACCTGCAGTTTACG-1 IGHV1-9;IGKV5-43;IGK.. None;None;None IGHJ4;IGKJ2;IGKJ1
#>                              c_gene      chains     isotype
#>                         <character> <character> <character>
#> 1_AAACCTGAGTAGGTGC-1           IGKC         IGK        None
#> 1_AAACCTGAGTCAAGCG-1      IGKC;IGKC     IGK;IGK        None
#> 1_AAACCTGCAGTTTACG-1 IGHM;IGKC;IGKC IGH;IGK;IGK        IGHM
#>                                        cdr3                cdr3_nt       reads
#>                                 <character>            <character> <character>
#> 1_AAACCTGAGTAGGTGC-1            CQQSNSWPYTF TGTCAACAGAGTAACAGCTG..        1054
#> 1_AAACCTGAGTCAAGCG-1 CQQYNSYPYTF;CLQHGESP.. TGTCAGCAATATAACAGCTA..     570;234
#> 1_AAACCTGCAGTTTACG-1 CATLWYDAMDYW;CQQSNSW.. TGTGCAACTCTATGGTACGA.. 144;460;156
#>                             umis     productive    full_length    paired
#>                      <character>    <character>    <character> <logical>
#> 1_AAACCTGAGTAGGTGC-1          70           TRUE           TRUE     FALSE
#> 1_AAACCTGAGTCAAGCG-1       31;17      TRUE;TRUE      TRUE;TRUE     FALSE
#> 1_AAACCTGCAGTTTACG-1     7;28;10 TRUE;TRUE;TRUE TRUE;TRUE;TRUE      TRUE
#>                      n_insertion  n_deletion  n_mismatch cdr3_length
#>                      <character> <character> <character> <character>
#> 1_AAACCTGAGTAGGTGC-1           0           4           2          11
#> 1_AAACCTGAGTCAAGCG-1         0;0         4;4         0;0       11;11
#> 1_AAACCTGCAGTTTACG-1       4;0;0       0;4;3       4;2;0    12;11;11
#>                      cdr3_nt_length  n_chains unique_chains
#>                         <character> <integer>   <character>
#> 1_AAACCTGAGTAGGTGC-1             33         1           IGK
#> 1_AAACCTGAGTCAAGCG-1          33;33         2           IGK
#> 1_AAACCTGCAGTTTACG-1       36;33;33         3       IGH_IGK

# Creating an index column to use for filtering/plotting
# this creates a column indicating which cells have no insertions
# the V(D)J data can be filtered based on this new column
res <- summarize_vdj(
  vdj_so,
  vdj_cols = "n_insertion",
  fn = ~ all(.x == 0),
  col_names = "no_insertions"
)

res <- filter_vdj(
  res,
  filt = no_insertions
)

head(res@meta.data, 3)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAACCTGAGTAGGTGC-1     avid_1          1            1               1
#> 1_AAACCTGAGTCAAGCG-1     avid_1          0            0               1
#> 1_AAACCTGCAGTTTACG-1     avid_1          4            4               0
#>                      seurat_clusters    UMAP_1     UMAP_2 clonotype_id
#> 1_AAACCTGAGTAGGTGC-1               1  2.561658 -0.6746224   clonotype1
#> 1_AAACCTGAGTCAAGCG-1               1 -3.402685  6.8385974 clonotype536
#> 1_AAACCTGCAGTTTACG-1               0  1.190111 -2.1084603         <NA>
#>                                   v_gene    d_gene      j_gene    c_gene
#> 1_AAACCTGAGTAGGTGC-1            IGKV5-43      None       IGKJ2      IGKC
#> 1_AAACCTGAGTCAAGCG-1 IGKV6-15;IGKV14-126 None;None IGKJ2;IGKJ2 IGKC;IGKC
#> 1_AAACCTGCAGTTTACG-1                <NA>      <NA>        <NA>      <NA>
#>                       chains isotype                    cdr3
#> 1_AAACCTGAGTAGGTGC-1     IGK    None             CQQSNSWPYTF
#> 1_AAACCTGAGTCAAGCG-1 IGK;IGK    None CQQYNSYPYTF;CLQHGESPYTF
#> 1_AAACCTGCAGTTTACG-1    <NA>    <NA>                    <NA>
#>                                                                                  cdr3_nt
#> 1_AAACCTGAGTAGGTGC-1                                   TGTCAACAGAGTAACAGCTGGCCGTACACGTTC
#> 1_AAACCTGAGTCAAGCG-1 TGTCAGCAATATAACAGCTATCCGTACACGTTC;TGTCTACAGCATGGTGAGAGCCCGTACACGTTC
#> 1_AAACCTGCAGTTTACG-1                                                                <NA>
#>                        reads  umis productive full_length paired n_insertion
#> 1_AAACCTGAGTAGGTGC-1    1054    70       TRUE        TRUE  FALSE           0
#> 1_AAACCTGAGTCAAGCG-1 570;234 31;17  TRUE;TRUE   TRUE;TRUE  FALSE         0;0
#> 1_AAACCTGCAGTTTACG-1    <NA>  <NA>       <NA>        <NA>     NA        <NA>
#>                      n_deletion n_mismatch cdr3_length cdr3_nt_length n_chains
#> 1_AAACCTGAGTAGGTGC-1          4          2          11             33        1
#> 1_AAACCTGAGTCAAGCG-1        4;4        0;0       11;11          33;33        2
#> 1_AAACCTGCAGTTTACG-1       <NA>       <NA>        <NA>           <NA>       NA
#>                      no_insertions
#> 1_AAACCTGAGTAGGTGC-1          TRUE
#> 1_AAACCTGAGTCAAGCG-1          TRUE
#> 1_AAACCTGCAGTTTACG-1            NA