Fetch V(D)J data from object

Fetch per-chain V(D)J data from object. Within the object meta.data, each row represents a single cell and can include information for multiple chains. This function can return a data.frame where each row represents a single chain. This is useful for plotting per-chain metrics such as CDR3 length or the number of insertions/deletions.

Usage

fetch_vdj(
  input,
  data_cols = NULL,
  clonotype_col = NULL,
  filter_cells = FALSE,
  per_cell = FALSE,
  unnest = TRUE,
  sep = ";"
)

Arguments

input

Single cell object or data.frame containing V(D)J data. If a data.frame is provided, the cell barcodes should be stored as row names.

data_cols

meta.data columns containing per-chain V(D)J data to unnest. If NULL, V(D)J data are automatically selected by identifying columns that have NAs in the same rows as clonotype_col.

clonotype_col

meta.data column containing clonotype IDs. This column is used to determine which columns have V(D)J data. If both clonotype_col and data_cols are NULL, all columns are included.

filter_cells

Remove cells that do not have V(D)J data, clonotype_col must be provided to determine which cells to filter.

per_cell

Return per-cell data instead of per-chain data.

unnest

If FALSE, a nested data.frame is returned where each row represents a cell and V(D)J data is stored as list-cols. If TRUE, columns are unnested so each row represents a single chain.

sep

Separator used for storing per cell V(D)J data. This is used to identify columns containing per-chain data that can be unnested.

Value

data.frame containing V(D)J data

Examples

# Fetch per-chain V(D)J data
fetch_vdj(vdj_so)
#> # A tibble: 258 × 28
#>    .cell_id  orig.…¹ nCoun…² nFeat…³ RNA_s…⁴ seura…⁵ UMAP_1 UMAP_2 clono…⁶
#>    <chr>     <chr>     <dbl>   <int> <fct>   <fct>    <dbl>  <dbl> <chr>  
#>  1 1_AAACCT… avid_1        1       1 1       1        2.56  -0.675 clonot…
#>  2 1_AAACCT… avid_1        0       0 1       1       -3.40   6.84  clonot…
#>  3 1_AAACCT… avid_1        0       0 1       1       -3.40   6.84  clonot…
#>  4 1_AAACCT… avid_1        4       4 0       0        1.19  -2.11  clonot…
#>  5 1_AAACCT… avid_1        4       4 0       0        1.19  -2.11  clonot…
#>  6 1_AAACCT… avid_1        4       4 0       0        1.19  -2.11  clonot…
#>  7 1_AAACCT… avid_1        9       8 0       0        0.137 -0.416 clonot…
#>  8 1_AACGTT… avid_1        1       1 1       1       -0.914 -1.03  NA     
#>  9 1_AATCGG… avid_1        1       1 1       1        3.09  -1.74  NA     
#> 10 1_ACATCA… avid_1        2       2 0       0        2.22  -1.39  NA     
#> # … with 248 more rows, 19 more variables: v_gene <chr>, d_gene <chr>,
#> #   j_gene <chr>, c_gene <chr>, chains <chr>, isotype <chr>, cdr3 <chr>,
#> #   cdr3_nt <chr>, reads <dbl>, umis <dbl>, productive <lgl>,
#> #   full_length <lgl>, paired <lgl>, n_insertion <dbl>, n_deletion <dbl>,
#> #   n_mismatch <dbl>, cdr3_length <dbl>, cdr3_nt_length <dbl>,
#> #   n_chains <int>, and abbreviated variable names ¹​orig.ident,
#> #   ²​nCount_RNA, ³​nFeature_RNA, ⁴​RNA_snn_res.0.3, ⁵​seurat_clusters, …

# To increase performance, specify which columns to return per-chain data,
# per-cell data will be returned for all other columns
fetch_vdj(
  vdj_sce,
  data_cols = c("chains", "n_insertion")
)
#> # A tibble: 258 × 28
#>    .cell_id  orig.…¹ nCoun…² nFeat…³ RNA_s…⁴ seura…⁵ UMAP_1 UMAP_2 clono…⁶
#>    <chr>     <chr>     <dbl>   <int> <fct>   <fct>    <dbl>  <dbl> <chr>  
#>  1 1_AAACCT… avid_1        1       1 1       1        2.56  -0.675 clonot…
#>  2 1_AAACCT… avid_1        0       0 1       1       -3.40   6.84  clonot…
#>  3 1_AAACCT… avid_1        0       0 1       1       -3.40   6.84  clonot…
#>  4 1_AAACCT… avid_1        4       4 0       0        1.19  -2.11  clonot…
#>  5 1_AAACCT… avid_1        4       4 0       0        1.19  -2.11  clonot…
#>  6 1_AAACCT… avid_1        4       4 0       0        1.19  -2.11  clonot…
#>  7 1_AAACCT… avid_1        9       8 0       0        0.137 -0.416 clonot…
#>  8 1_AACGTT… avid_1        1       1 1       1       -0.914 -1.03  NA     
#>  9 1_AATCGG… avid_1        1       1 1       1        3.09  -1.74  NA     
#> 10 1_ACATCA… avid_1        2       2 0       0        2.22  -1.39  NA     
#> # … with 248 more rows, 19 more variables: v_gene <chr>, d_gene <chr>,
#> #   j_gene <chr>, c_gene <chr>, chains <chr>, isotype <chr>, cdr3 <chr>,
#> #   cdr3_nt <chr>, reads <chr>, umis <chr>, productive <chr>,
#> #   full_length <chr>, paired <lgl>, n_insertion <dbl>, n_deletion <chr>,
#> #   n_mismatch <chr>, cdr3_length <chr>, cdr3_nt_length <chr>,
#> #   n_chains <int>, and abbreviated variable names ¹​orig.ident,
#> #   ²​nCount_RNA, ³​nFeature_RNA, ⁴​RNA_snn_res.0.3, ⁵​seurat_clusters, …

# Only include cells that have V(D)J data
# clonotype_col must be specified to identify cells with V(D)J data
fetch_vdj(
  vdj_so,
  filter_cells = TRUE,
  clonotype_col = "clonotype_id"
)
#> # A tibble: 143 × 28
#>    .cell_id  orig.…¹ nCoun…² nFeat…³ RNA_s…⁴ seura…⁵ UMAP_1 UMAP_2 clono…⁶
#>    <chr>     <chr>     <dbl>   <int> <fct>   <fct>    <dbl>  <dbl> <chr>  
#>  1 1_AAACCT… avid_1        1       1 1       1        2.56  -0.675 clonot…
#>  2 1_AAACCT… avid_1        0       0 1       1       -3.40   6.84  clonot…
#>  3 1_AAACCT… avid_1        0       0 1       1       -3.40   6.84  clonot…
#>  4 1_AAACCT… avid_1        4       4 0       0        1.19  -2.11  clonot…
#>  5 1_AAACCT… avid_1        4       4 0       0        1.19  -2.11  clonot…
#>  6 1_AAACCT… avid_1        4       4 0       0        1.19  -2.11  clonot…
#>  7 1_AAACCT… avid_1        9       8 0       0        0.137 -0.416 clonot…
#>  8 1_ACCTTT… avid_1        4       4 0       0        1.13  -2.42  clonot…
#>  9 1_ACCTTT… avid_1        4       4 0       0        1.13  -2.42  clonot…
#> 10 1_ACCTTT… avid_1        4       4 0       0        1.13  -2.42  clonot…
#> # … with 133 more rows, 19 more variables: v_gene <chr>, d_gene <chr>,
#> #   j_gene <chr>, c_gene <chr>, chains <chr>, isotype <chr>, cdr3 <chr>,
#> #   cdr3_nt <chr>, reads <dbl>, umis <dbl>, productive <lgl>,
#> #   full_length <lgl>, paired <lgl>, n_insertion <dbl>, n_deletion <dbl>,
#> #   n_mismatch <dbl>, cdr3_length <dbl>, cdr3_nt_length <dbl>,
#> #   n_chains <int>, and abbreviated variable names ¹​orig.ident,
#> #   ²​nCount_RNA, ³​nFeature_RNA, ⁴​RNA_snn_res.0.3, ⁵​seurat_clusters, …