Import V(D)J data

Import V(D)J data

Usage

import_vdj(
  input = NULL,
  vdj_dir = NULL,
  prefix = "",
  data_cols = NULL,
  filter_chains = TRUE,
  filter_paired = FALSE,
  define_clonotypes = NULL,
  include_mutations = FALSE,
  aggr_dir = NULL,
  sep = ";"
)

Arguments

input

Object containing single cell data, if set to NULL a data.frame containing V(D)J results will be returned

vdj_dir

Directory containing the output from cellranger vdj. A vector or named vector can be given to load data from multiple runs. If a named vector is given, the cell barcodes will be prefixed with the provided names. This mimics the behavior of Seurat::Read10X().

prefix

Prefix to add to new columns

data_cols

Additional columns from filtered_contig_annotations.csv to include in object.

filter_chains

Only include chains with at least one productive and full length contig.

filter_paired

Only include clonotypes with paired chains. For TCR data each clonotype must have at least one TRA and TRB chain, for BCR data each clonotype must have at least one IGH chain and at least one IGK or IGL chain.

define_clonotypes

Define clonotype IDs based on V(D)J data. This is useful if the V(D)J datasets being loaded do not have consistent clonotype IDs, i.e., clonotype1 is not the same across samples. Possible values are:

'cdr3aa', define clonotypes based on the CDR3 amino acid sequence

'cdr3nt', define clonotypes based on the CDR3 nucleotide sequence

'cdr3_gene', define clonotypes based on the combination of the CDR3 nucleotide sequence and the V(D)J genes.

When defining clonotypes, only productive full length chains will be used. Set to NULL (default) to use the clonotype IDs already present in the input data.

include_mutations

Include information about the number of insertions/deletions/mismatches for each chain. This requires the concat_ref.bam file from cellranger vdj to be present the directory provided to vdj_dir. If include_mutations is TRUE, filter_chains is also automatically set TRUE since indel data is only available for productive chains.

aggr_dir

Path to cellranger aggr output. To include mutation information for each chain, also provide paths to the original cellranger vdj output directories using the vdj_dir argument.

To correctly match cell barcodes to those in the object, gene expression data for each sample must be loaded in the same order as the samples were specified in the cellranger aggr config file. In addition, if loading mutation data, sample paths provided to the vdj_dir argument must also be in the same order as the samples were specified in the cellranger aggr config file.

sep

Separator to use for storing per cell V(D)J data

Value

Single cell object or data.frame with added V(D)J data

Examples

# Loading multiple datasets
vdj_dir <- c(
  system.file("extdata/bcr_1/outs", package = "djvdj"),
  system.file("extdata/bcr_2/outs", package = "djvdj")
)

vdj_so <- import_vdj(tiny_so, vdj_dir, include_mutations = FALSE)

head(vdj_so@meta.data, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAGCCGCAGCTTATCG-1     avid_1          0            0               0
#>                      seurat_clusters    UMAP_1    UMAP_2 clonotype_id
#> 1_AAGCCGCAGCTTATCG-1               0 -5.977054 -2.418108         <NA>
#>                      exact_subclonotype_id chains n_chains cdr3 cdr3_nt
#> 1_AAGCCGCAGCTTATCG-1                    NA   <NA>       NA <NA>    <NA>
#>                      cdr3_length cdr3_nt_length v_gene d_gene j_gene
#> 1_AAGCCGCAGCTTATCG-1        <NA>           <NA>   <NA>   <NA>   <NA>
#>                      c_gene isotype reads umis productive full_length
#> 1_AAGCCGCAGCTTATCG-1   <NA>    <NA>  <NA> <NA>       <NA>        <NA>
#>                      paired
#> 1_AAGCCGCAGCTTATCG-1     NA

# Specifying cell prefixes
# if cell prefixes are not specified when loading multiple datasets,
# prefixes will be automatically generated in a similar manner as
# Seurat::Read10X
vdj_so <- import_vdj(
  tiny_so,
  vdj_dir = vdj_dir,
  include_mutations = FALSE
)

head(vdj_so@meta.data, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAGCCGCAGCTTATCG-1     avid_1          0            0               0
#>                      seurat_clusters    UMAP_1    UMAP_2 clonotype_id
#> 1_AAGCCGCAGCTTATCG-1               0 -5.977054 -2.418108         <NA>
#>                      exact_subclonotype_id chains n_chains cdr3 cdr3_nt
#> 1_AAGCCGCAGCTTATCG-1                    NA   <NA>       NA <NA>    <NA>
#>                      cdr3_length cdr3_nt_length v_gene d_gene j_gene
#> 1_AAGCCGCAGCTTATCG-1        <NA>           <NA>   <NA>   <NA>   <NA>
#>                      c_gene isotype reads umis productive full_length
#> 1_AAGCCGCAGCTTATCG-1   <NA>    <NA>  <NA> <NA>       <NA>        <NA>
#>                      paired
#> 1_AAGCCGCAGCTTATCG-1     NA

# Specifying cell prefixes using vector names
# if a named vector is passed, the names will be used as the cell prefixes
vdj_dir <- c(
  "1" = system.file("extdata/bcr_1/outs", package = "djvdj"),
  "2" = system.file("extdata/bcr_2/outs", package = "djvdj")
)

vdj_so <- import_vdj(tiny_so, vdj_dir, include_mutations = FALSE)

head(vdj_so@meta.data, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAGCCGCAGCTTATCG-1     avid_1          0            0               0
#>                      seurat_clusters    UMAP_1    UMAP_2 clonotype_id
#> 1_AAGCCGCAGCTTATCG-1               0 -5.977054 -2.418108         <NA>
#>                      exact_subclonotype_id chains n_chains cdr3 cdr3_nt
#> 1_AAGCCGCAGCTTATCG-1                    NA   <NA>       NA <NA>    <NA>
#>                      cdr3_length cdr3_nt_length v_gene d_gene j_gene
#> 1_AAGCCGCAGCTTATCG-1        <NA>           <NA>   <NA>   <NA>   <NA>
#>                      c_gene isotype reads umis productive full_length
#> 1_AAGCCGCAGCTTATCG-1   <NA>    <NA>  <NA> <NA>       <NA>        <NA>
#>                      paired
#> 1_AAGCCGCAGCTTATCG-1     NA

# Only include V(D)J data for productive full length chains
vdj_so <- import_vdj(
  tiny_so,
  vdj_dir = vdj_dir,
  filter_chains = TRUE,
  include_mutations = FALSE
)

head(vdj_so@meta.data, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAGCCGCAGCTTATCG-1     avid_1          0            0               0
#>                      seurat_clusters    UMAP_1    UMAP_2 clonotype_id
#> 1_AAGCCGCAGCTTATCG-1               0 -5.977054 -2.418108         <NA>
#>                      exact_subclonotype_id chains n_chains cdr3 cdr3_nt
#> 1_AAGCCGCAGCTTATCG-1                    NA   <NA>       NA <NA>    <NA>
#>                      cdr3_length cdr3_nt_length v_gene d_gene j_gene
#> 1_AAGCCGCAGCTTATCG-1        <NA>           <NA>   <NA>   <NA>   <NA>
#>                      c_gene isotype reads umis productive full_length
#> 1_AAGCCGCAGCTTATCG-1   <NA>    <NA>  <NA> <NA>       <NA>        <NA>
#>                      paired
#> 1_AAGCCGCAGCTTATCG-1     NA

# Only include V(D)J data for cells with paired chains
vdj_so <- import_vdj(
  tiny_so,
  vdj_dir = vdj_dir,
  filter_paired = TRUE,
  include_mutations = FALSE
)

head(vdj_so@meta.data, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAGCCGCAGCTTATCG-1     avid_1          0            0               0
#>                      seurat_clusters    UMAP_1    UMAP_2 clonotype_id
#> 1_AAGCCGCAGCTTATCG-1               0 -5.977054 -2.418108         <NA>
#>                      exact_subclonotype_id chains n_chains cdr3 cdr3_nt
#> 1_AAGCCGCAGCTTATCG-1                    NA   <NA>       NA <NA>    <NA>
#>                      cdr3_length cdr3_nt_length v_gene d_gene j_gene
#> 1_AAGCCGCAGCTTATCG-1        <NA>           <NA>   <NA>   <NA>   <NA>
#>                      c_gene isotype reads umis productive full_length
#> 1_AAGCCGCAGCTTATCG-1   <NA>    <NA>  <NA> <NA>       <NA>        <NA>
#>                      paired
#> 1_AAGCCGCAGCTTATCG-1     NA

# Defining clonotypes
# this is useful if the original clonotype IDs are not consistent across
# datasets, i.e. clonotype1 is not the same for all samples
vdj_so <- import_vdj(
  tiny_so,
  vdj_dir = vdj_dir,
  define_clonotypes = "cdr3_gene",
  include_mutations = FALSE
)

head(vdj_so@meta.data, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAGCCGCAGCTTATCG-1     avid_1          0            0               0
#>                      seurat_clusters    UMAP_1    UMAP_2
#> 1_AAGCCGCAGCTTATCG-1               0 -5.977054 -2.418108
#>                      exact_subclonotype_id chains n_chains cdr3 cdr3_nt
#> 1_AAGCCGCAGCTTATCG-1                    NA   <NA>       NA <NA>    <NA>
#>                      cdr3_length cdr3_nt_length v_gene d_gene j_gene
#> 1_AAGCCGCAGCTTATCG-1        <NA>           <NA>   <NA>   <NA>   <NA>
#>                      c_gene isotype reads umis productive full_length
#> 1_AAGCCGCAGCTTATCG-1   <NA>    <NA>  <NA> <NA>       <NA>        <NA>
#>                      paired clonotype_id
#> 1_AAGCCGCAGCTTATCG-1     NA         <NA>

# Include mutation information for each chain
# this information will be included if the file concat_ref.bam is present
# including mutation information will cause data import to be slower
vdj_so <- import_vdj(
  tiny_so,
  vdj_dir = vdj_dir,
  include_mutations = TRUE
)
#> Warning: airr_rearrangement.tsv not found in /home/runner/work/_temp/Library/djvdj/extdata/bcr_1/outs.
#> Warning: airr_rearrangement.tsv not found in /home/runner/work/_temp/Library/djvdj/extdata/bcr_2/outs.
#> Warning: To add mutation data to object, concat_ref.bam and airr_rearrangement.tsv must be present for all samples, check that these files are in the provided directory paths, mutation data not added to object.

head(vdj_so@meta.data, 1)
#>                      orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#> 1_AAGCCGCAGCTTATCG-1     avid_1          0            0               0
#>                      seurat_clusters    UMAP_1    UMAP_2 clonotype_id
#> 1_AAGCCGCAGCTTATCG-1               0 -5.977054 -2.418108         <NA>
#>                      exact_subclonotype_id chains n_chains cdr3 cdr3_nt
#> 1_AAGCCGCAGCTTATCG-1                    NA   <NA>       NA <NA>    <NA>
#>                      cdr3_length cdr3_nt_length v_gene d_gene j_gene
#> 1_AAGCCGCAGCTTATCG-1        <NA>           <NA>   <NA>   <NA>   <NA>
#>                      c_gene isotype reads umis productive full_length
#> 1_AAGCCGCAGCTTATCG-1   <NA>    <NA>  <NA> <NA>       <NA>        <NA>
#>                      paired
#> 1_AAGCCGCAGCTTATCG-1     NA

# Using import_vdj outside of Seurat
# SingleCellExperiment objects are also compatible, or if an input object is
# omitted, a data.frame containing the V(D)J data will be returned
vdj_sce <- import_vdj(tiny_sce, vdj_dir)

head(vdj_sce@colData, 1)
#> DataFrame with 1 row and 25 columns
#>                       orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.3
#>                      <character>  <numeric>    <integer>        <factor>
#> 1_AAGCCGCAGCTTATCG-1      avid_1          0            0               0
#>                      seurat_clusters    UMAP_1    UMAP_2 clonotype_id
#>                             <factor> <numeric> <numeric>  <character>
#> 1_AAGCCGCAGCTTATCG-1               0  -5.97705  -2.41811           NA
#>                      exact_subclonotype_id      chains  n_chains
#>                                  <numeric> <character> <integer>
#> 1_AAGCCGCAGCTTATCG-1                    NA          NA        NA
#>                             cdr3     cdr3_nt cdr3_length cdr3_nt_length
#>                      <character> <character> <character>    <character>
#> 1_AAGCCGCAGCTTATCG-1          NA          NA          NA             NA
#>                           v_gene      d_gene      j_gene      c_gene
#>                      <character> <character> <character> <character>
#> 1_AAGCCGCAGCTTATCG-1          NA          NA          NA          NA
#>                          isotype       reads        umis  productive
#>                      <character> <character> <character> <character>
#> 1_AAGCCGCAGCTTATCG-1          NA          NA          NA          NA
#>                      full_length    paired
#>                      <character> <logical>
#> 1_AAGCCGCAGCTTATCG-1          NA        NA

vdj_df <- import_vdj(vdj_dir = vdj_dir, include_mutations = FALSE)

head(vdj_df, 1)
#>                     clonotype_id exact_subclonotype_id chains n_chains
#> 1ACGGAGACATGCTGGC-1   clonotype1                     1    IGK        1
#>                            cdr3                           cdr3_nt
#> 1ACGGAGACATGCTGGC-1 CQQSNSWPYTF TGTCAACAGAGTAACAGCTGGCCGTACACGTTC
#>                     cdr3_length cdr3_nt_length   v_gene d_gene j_gene
#> 1ACGGAGACATGCTGGC-1          11             33 IGKV5-43   None  IGKJ2
#>                     c_gene isotype reads umis productive full_length
#> 1ACGGAGACATGCTGGC-1   IGKC    None   394   24       TRUE        TRUE
#>                     paired
#> 1ACGGAGACATGCTGGC-1  FALSE