Add V(D)J data to Seurat object

Add V(D)J data to Seurat object

import_vdj(
  sobj_in,
  vdj_dir,
  prefix = "",
  cell_prefix = "",
  filter_contigs = TRUE,
  sep = ";"
)

Arguments

sobj_in
Seurat object

vdj_dir
cellranger vdj output directory

prefix
Prefix to add to new meta.data columns

cell_prefix
Prefix to add to cell barcodes

filter_contigs
Only include chains with at least one productive contig

sep
Separator to use for storing per cell clonotype information in the meta.data

Value

Seurat object with V(D)J data added to meta.data